refactor: home page component
