+++ assets/js/wepos-react.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";var e={69:(e,t)=>{Object.prototype.toString},338:(e,t,a)=>{var n=a(795);t.H=n.createRoot,n.hydrateRoot},795:e=>{e.exports=window.ReactDOM}},t={};function a(n){var r=t[n];if(void 0!==r)return r.exports;var l=t[n]={exports:{}};return e[n](l,l.exports,a),l.exports}a.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return a.d(t,{a:t}),t},a.d=(e,t)=>{for(var n in t)a.o(t,n)&&!a.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},a.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t);const n=window.wp.element,r=window.React;var l=a.n(r),o=a(338),i=(a(69),"popstate");function s(e={}){return function(e,t,a,n={}){let{window:r=document.defaultView,v5Compat:l=!1}=n,o=r.history,s="POP",u=null,h=f();function f(){return(o.state||{idx:null}).idx}function y(){s="POP";let e=f(),t=null==e?null:e-h;h=e,u&&u({action:s,location:E.location,delta:t})}function g(e){return function(e,t=!1){let a="http://localhost";"undefined"!=typeof window&&(a="null"!==window.location.origin?window.location.origin:window.location.href),c(a,"No window.location.(origin|href) available to create URL");let n="string"==typeof e?e:p(e);return n=n.replace(/ $/,"%20"),!t&&n.startsWith("//")&&(n=a+n),new URL(n,a)}(e)}null==h&&(h=0,o.replaceState({...o.state,idx:h},""));let E={get action(){return s},get location(){return e(r,o)},listen(e){if(u)throw new Error("A history only accepts one active listener");return r.addEventListener(i,y),u=e,()=>{r.removeEventListener(i,y),u=null}},createHref:e=>t(r,e),createURL:g,encodeLocation(e){let t=g(e);return{pathname:t.pathname,search:t.search,hash:t.hash}},push:function(e,t){s="PUSH";let n=d(E.location,e,t);a&&a(n,e),h=f()+1;let i=m(n,h),c=E.createHref(n);try{o.pushState(i,"",c)}catch(e){if(e instanceof DOMException&&"DataCloneError"===e.name)throw e;r.location.assign(c)}l&&u&&u({action:s,location:E.location,delta:1})},replace:function(e,t){s="REPLACE";let n=d(E.location,e,t);a&&a(n,e),h=f();let r=m(n,h),i=E.createHref(n);o.replaceState(r,"",i),l&&u&&u({action:s,location:E.location,delta:0})},go:e=>o.go(e)};return E}((function(e,t){let{pathname:a="/",search:n="",hash:r=""}=h(e.location.hash.substring(1));return a.startsWith("/")||a.startsWith(".")||(a="/"+a),d("",{pathname:a,search:n,hash:r},t.state&&t.state.usr||null,t.state&&t.state.key||"default")}),(function(e,t){let a=e.document.querySelector("base"),n="";if(a&&a.getAttribute("href")){let t=e.location.href,a=t.indexOf("#");n=-1===a?t:t.slice(0,a)}return n+"#"+("string"==typeof t?t:p(t))}),(function(e,t){u("/"===e.pathname.charAt(0),`relative pathnames are not supported in hash history.push(${JSON.stringify(t)})`)}),e)}function c(e,t){if(!1===e||null==e)throw new Error(t)}function u(e,t){if(!e){"undefined"!=typeof console&&console.warn(t);try{throw new Error(t)}catch(e){}}}function m(e,t){return{usr:e.state,key:e.key,idx:t}}function d(e,t,a=null,n){return{pathname:"string"==typeof e?e:e.pathname,search:"",hash:"",..."string"==typeof t?h(t):t,state:a,key:t&&t.key||n||Math.random().toString(36).substring(2,10)}}function p({pathname:e="/",search:t="",hash:a=""}){return t&&"?"!==t&&(e+="?"===t.charAt(0)?t:"?"+t),a&&"#"!==a&&(e+="#"===a.charAt(0)?a:"#"+a),e}function h(e){let t={};if(e){let a=e.indexOf("#");a>=0&&(t.hash=e.substring(a),e=e.substring(0,a));let n=e.indexOf("?");n>=0&&(t.search=e.substring(n),e=e.substring(0,n)),e&&(t.pathname=e)}return t}function f(e,t,a="/"){return function(e,t,a,n){let r=$(("string"==typeof t?h(t):t).pathname||"/",a);if(null==r)return null;let l=y(e);!function(e){e.sort(((e,t)=>e.score!==t.score?t.score-e.score:function(e,t){return e.length===t.length&&e.slice(0,-1).every(((e,a)=>e===t[a]))?e[e.length-1]-t[t.length-1]:0}(e.routesMeta.map((e=>e.childrenIndex)),t.routesMeta.map((e=>e.childrenIndex)))))}(l);let o=null;for(let e=0;null==o&&e<l.length;++e){let t=R(r);o=x(l[e],t,n)}return o}(e,t,a,!1)}function y(e,t=[],a=[],n=""){let r=(e,r,l)=>{let o={relativePath:void 0===l?e.path||"":l,caseSensitive:!0===e.caseSensitive,childrenIndex:r,route:e};o.relativePath.startsWith("/")&&(c(o.relativePath.startsWith(n),`Absolute route path "${o.relativePath}" nested under path "${n}" is not valid. An absolute child route path must start with the combined path of all its parent routes.`),o.relativePath=o.relativePath.slice(n.length));let i=F([n,o.relativePath]),s=a.concat(o);e.children&&e.children.length>0&&(c(!0!==e.index,`Index routes must not have child routes. Please remove all child routes from route path "${i}".`),y(e.children,t,s,i)),(null!=e.path||e.index)&&t.push({path:i,score:C(i,e.index),routesMeta:s})};return e.forEach(((e,t)=>{if(""!==e.path&&e.path?.includes("?"))for(let a of g(e.path))r(e,t,a);else r(e,t)})),t}function g(e){let t=e.split("/");if(0===t.length)return[];let[a,...n]=t,r=a.endsWith("?"),l=a.replace(/\?$/,"");if(0===n.length)return r?[l,""]:[l];let o=g(n.join("/")),i=[];return i.push(...o.map((e=>""===e?l:[l,e].join("/")))),r&&i.push(...o),i.map((t=>e.startsWith("/")&&""===t?"/":t))}new WeakMap;var E=/^:[\w-]+$/,v=3,w=2,b=1,N=10,_=-2,S=e=>"*"===e;function C(e,t){let a=e.split("/"),n=a.length;return a.some(S)&&(n+=_),t&&(n+=w),a.filter((e=>!S(e))).reduce(((e,t)=>e+(E.test(t)?v:""===t?b:N)),n)}function x(e,t,a=!1){let{routesMeta:n}=e,r={},l="/",o=[];for(let e=0;e<n.length;++e){let i=n[e],s=e===n.length-1,c="/"===l?t:t.slice(l.length)||"/",u=k({path:i.relativePath,caseSensitive:i.caseSensitive,end:s},c),m=i.route;if(!u&&s&&a&&!n[n.length-1].route.index&&(u=k({path:i.relativePath,caseSensitive:i.caseSensitive,end:!1},c)),!u)return null;Object.assign(r,u.params),o.push({params:r,pathname:F([l,u.pathname]),pathnameBase:O(F([l,u.pathnameBase])),route:m}),"/"!==u.pathnameBase&&(l=F([l,u.pathnameBase]))}return o}function k(e,t){"string"==typeof e&&(e={path:e,caseSensitive:!1,end:!0});let[a,n]=function(e,t=!1,a=!0){u("*"===e||!e.endsWith("*")||e.endsWith("/*"),`Route path "${e}" will be treated as if it were "${e.replace(/\*$/,"/*")}" because the \`*\` character must always follow a \`/\` in the pattern. To get rid of this warning, please change the route path to "${e.replace(/\*$/,"/*")}".`);let n=[],r="^"+e.replace(/\/*\*?$/,"").replace(/^\/*/,"/").replace(/[\\.*+^${}|()[\]]/g,"\\$&").replace(/\/:([\w-]+)(\?)?/g,((e,t,a)=>(n.push({paramName:t,isOptional:null!=a}),a?"/?([^\\/]+)?":"/([^\\/]+)")));return e.endsWith("*")?(n.push({paramName:"*"}),r+="*"===e||"/*"===e?"(.*)$":"(?:\\/(.+)|\\/*)$"):a?r+="\\/*$":""!==e&&"/"!==e&&(r+="(?:(?=\\/|$))"),[new RegExp(r,t?void 0:"i"),n]}(e.path,e.caseSensitive,e.end),r=t.match(a);if(!r)return null;let l=r[0],o=l.replace(/(.)\/+$/,"$1"),i=r.slice(1);return{params:n.reduce(((e,{paramName:t,isOptional:a},n)=>{if("*"===t){let e=i[n]||"";o=l.slice(0,l.length-e.length).replace(/(.)\/+$/,"$1")}const r=i[n];return e[t]=a&&!r?void 0:(r||"").replace(/%2F/g,"/"),e}),{}),pathname:l,pathnameBase:o,pattern:e}}function R(e){try{return e.split("/").map((e=>decodeURIComponent(e).replace(/\//g,"%2F"))).join("/")}catch(t){return u(!1,`The URL path "${e}" could not be decoded because it is a malformed URL segment. This is probably due to a bad percent encoding (${t}).`),e}}function $(e,t){if("/"===t)return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let a=t.endsWith("/")?t.length-1:t.length,n=e.charAt(a);return n&&"/"!==n?null:e.slice(a)||"/"}function P(e,t,a,n){return`Cannot include a '${e}' character in a manually specified \`to.${t}\` field [${JSON.stringify(n)}].  Please separate it out to the \`to.${a}\` field. Alternatively you may provide the full path as a string in <Link to="..."> and the router will parse it for you.`}function L(e){let t=function(e){return e.filter(((e,t)=>0===t||e.route.path&&e.route.path.length>0))}(e);return t.map(((e,a)=>a===t.length-1?e.pathname:e.pathnameBase))}function T(e,t,a,n=!1){let r;"string"==typeof e?r=h(e):(r={...e},c(!r.pathname||!r.pathname.includes("?"),P("?","pathname","search",r)),c(!r.pathname||!r.pathname.includes("#"),P("#","pathname","hash",r)),c(!r.search||!r.search.includes("#"),P("#","search","hash",r)));let l,o=""===e||""===r.pathname,i=o?"/":r.pathname;if(null==i)l=a;else{let e=t.length-1;if(!n&&i.startsWith("..")){let t=i.split("/");for(;".."===t[0];)t.shift(),e-=1;r.pathname=t.join("/")}l=e>=0?t[e]:"/"}let s=function(e,t="/"){let{pathname:a,search:n="",hash:r=""}="string"==typeof e?h(e):e,l=a?a.startsWith("/")?a:function(e,t){let a=t.replace(/\/+$/,"").split("/");return e.split("/").forEach((e=>{".."===e?a.length>1&&a.pop():"."!==e&&a.push(e)})),a.length>1?a.join("/"):"/"}(a,t):t;return{pathname:l,search:D(n),hash:M(r)}}(r,l),u=i&&"/"!==i&&i.endsWith("/"),m=(o||"."===i)&&a.endsWith("/");return s.pathname.endsWith("/")||!u&&!m||(s.pathname+="/"),s}var F=e=>e.join("/").replace(/\/\/+/g,"/"),O=e=>e.replace(/\/+$/,"").replace(/^\/*/,"/"),D=e=>e&&"?"!==e?e.startsWith("?")?e:"?"+e:"",M=e=>e&&"#"!==e?e.startsWith("#")?e:"#"+e:"";function A(e){return null!=e&&"number"==typeof e.status&&"string"==typeof e.statusText&&"boolean"==typeof e.internal&&"data"in e}var W=["POST","PUT","PATCH","DELETE"],q=(new Set(W),["GET",...W]);new Set(q),Symbol("ResetLoaderData");var B=r.createContext(null);B.displayName="DataRouter";var I=r.createContext(null);I.displayName="DataRouterState";var U=r.createContext({isTransitioning:!1});U.displayName="ViewTransition",r.createContext(new Map).displayName="Fetchers",r.createContext(null).displayName="Await";var j=r.createContext(null);j.displayName="Navigation";var H=r.createContext(null);H.displayName="Location";var J=r.createContext({outlet:null,matches:[],isDataRoute:!1});J.displayName="Route";var K=r.createContext(null);function z(){return null!=r.useContext(H)}function Y(){return c(z(),"useLocation() may be used only in the context of a <Router> component."),r.useContext(H).location}K.displayName="RouteError";var Q="You should call navigate() in a React.useEffect(), not when your component is first rendered.";function V(e){r.useContext(j).static||r.useLayoutEffect(e)}function G(){let{isDataRoute:e}=r.useContext(J);return e?function(){let{router:e}=function(e){let t=r.useContext(B);return c(t,re(e)),t}("useNavigate"),t=le("useNavigate"),a=r.useRef(!1);return V((()=>{a.current=!0})),r.useCallback((async(n,r={})=>{u(a.current,Q),a.current&&("number"==typeof n?e.navigate(n):await e.navigate(n,{fromRouteId:t,...r}))}),[e,t])}():function(){c(z(),"useNavigate() may be used only in the context of a <Router> component.");let e=r.useContext(B),{basename:t,navigator:a}=r.useContext(j),{matches:n}=r.useContext(J),{pathname:l}=Y(),o=JSON.stringify(L(n)),i=r.useRef(!1);return V((()=>{i.current=!0})),r.useCallback(((n,r={})=>{if(u(i.current,Q),!i.current)return;if("number"==typeof n)return void a.go(n);let s=T(n,JSON.parse(o),l,"path"===r.relative);null==e&&"/"!==t&&(s.pathname="/"===s.pathname?t:F([t,s.pathname])),(r.replace?a.replace:a.push)(s,r.state,r)}),[t,a,o,l,e])}()}function X(e,{relative:t}={}){let{matches:a}=r.useContext(J),{pathname:n}=Y(),l=JSON.stringify(L(a));return r.useMemo((()=>T(e,JSON.parse(l),n,"path"===t)),[e,l,n,t])}function Z(e,t,a,n){c(z(),"useRoutes() may be used only in the context of a <Router> component.");let{navigator:l}=r.useContext(j),{matches:o}=r.useContext(J),i=o[o.length-1],s=i?i.params:{},m=i?i.pathname:"/",d=i?i.pathnameBase:"/",p=i&&i.route;{let e=p&&p.path||"";ie(m,!p||e.endsWith("*")||e.endsWith("*?"),`You rendered descendant <Routes> (or called \`useRoutes()\`) at "${m}" (under <Route path="${e}">) but the parent route path has no trailing "*". This means if you navigate deeper, the parent won't match anymore and therefore the child routes will never render.\n\nPlease change the parent <Route path="${e}"> to <Route path="${"/"===e?"*":`${e}/*`}">.`)}let y,g=Y();if(t){let e="string"==typeof t?h(t):t;c("/"===d||e.pathname?.startsWith(d),`When overriding the location using \`<Routes location>\` or \`useRoutes(routes, location)\`, the location pathname must begin with the portion of the URL pathname that was matched by all parent routes. The current pathname base is "${d}" but pathname "${e.pathname}" was given in the \`location\` prop.`),y=e}else y=g;let E=y.pathname||"/",v=E;if("/"!==d){let e=d.replace(/^\//,"").split("/");v="/"+E.replace(/^\//,"").split("/").slice(e.length).join("/")}let w=f(e,{pathname:v});u(p||null!=w,`No routes matched location "${y.pathname}${y.search}${y.hash}" `),u(null==w||void 0!==w[w.length-1].route.element||void 0!==w[w.length-1].route.Component||void 0!==w[w.length-1].route.lazy,`Matched leaf route at location "${y.pathname}${y.search}${y.hash}" does not have an element or Component. This means it will render an <Outlet /> with a null value by default resulting in an "empty" page.`);let b=function(e,t=[],a=null){if(null==e){if(!a)return null;if(a.errors)e=a.matches;else{if(0!==t.length||a.initialized||!(a.matches.length>0))return null;e=a.matches}}let n=e,l=a?.errors;if(null!=l){let e=n.findIndex((e=>e.route.id&&void 0!==l?.[e.route.id]));c(e>=0,`Could not find a matching route for errors on route IDs: ${Object.keys(l).join(",")}`),n=n.slice(0,Math.min(n.length,e+1))}let o=!1,i=-1;if(a)for(let e=0;e<n.length;e++){let t=n[e];if((t.route.HydrateFallback||t.route.hydrateFallbackElement)&&(i=e),t.route.id){let{loaderData:e,errors:r}=a,l=t.route.loader&&!e.hasOwnProperty(t.route.id)&&(!r||void 0===r[t.route.id]);if(t.route.lazy||l){o=!0,n=i>=0?n.slice(0,i+1):[n[0]];break}}}return n.reduceRight(((e,s,c)=>{let u,m=!1,d=null,p=null;a&&(u=l&&s.route.id?l[s.route.id]:void 0,d=s.route.errorElement||te,o&&(i<0&&0===c?(ie("route-fallback",!1,"No `HydrateFallback` element provided to render during initial hydration"),m=!0,p=null):i===c&&(m=!0,p=s.route.hydrateFallbackElement||null)));let h=t.concat(n.slice(0,c+1)),f=()=>{let t;return t=u?d:m?p:s.route.Component?r.createElement(s.route.Component,null):s.route.element?s.route.element:e,r.createElement(ne,{match:s,routeContext:{outlet:e,matches:h,isDataRoute:null!=a},children:t})};return a&&(s.route.ErrorBoundary||s.route.errorElement||0===c)?r.createElement(ae,{location:a.location,revalidation:a.revalidation,component:d,error:u,children:f(),routeContext:{outlet:null,matches:h,isDataRoute:!0}}):f()}),null)}(w&&w.map((e=>Object.assign({},e,{params:Object.assign({},s,e.params),pathname:F([d,l.encodeLocation?l.encodeLocation(e.pathname).pathname:e.pathname]),pathnameBase:"/"===e.pathnameBase?d:F([d,l.encodeLocation?l.encodeLocation(e.pathnameBase).pathname:e.pathnameBase])}))),o,a,n);return t&&b?r.createElement(H.Provider,{value:{location:{pathname:"/",search:"",hash:"",state:null,key:"default",...y},navigationType:"POP"}},b):b}function ee(){let e=function(){let e=r.useContext(K),t=function(e){let t=r.useContext(I);return c(t,re(e)),t}("useRouteError"),a=le("useRouteError");return void 0!==e?e:t.errors?.[a]}(),t=A(e)?`${e.status} ${e.statusText}`:e instanceof Error?e.message:JSON.stringify(e),a=e instanceof Error?e.stack:null,n="rgba(200,200,200, 0.5)",l={padding:"0.5rem",backgroundColor:n},o={padding:"2px 4px",backgroundColor:n},i=null;return console.error("Error handled by React Router default ErrorBoundary:",e),i=r.createElement(r.Fragment,null,r.createElement("p",null,"💿 Hey developer 👋"),r.createElement("p",null,"You can provide a way better UX than this when your app throws errors by providing your own ",r.createElement("code",{style:o},"ErrorBoundary")," or"," ",r.createElement("code",{style:o},"errorElement")," prop on your route.")),r.createElement(r.Fragment,null,r.createElement("h2",null,"Unexpected Application Error!"),r.createElement("h3",{style:{fontStyle:"italic"}},t),a?r.createElement("pre",{style:l},a):null,i)}r.createContext(null);var te=r.createElement(ee,null),ae=class extends r.Component{constructor(e){super(e),this.state={location:e.location,revalidation:e.revalidation,error:e.error}}static getDerivedStateFromError(e){return{error:e}}static getDerivedStateFromProps(e,t){return t.location!==e.location||"idle"!==t.revalidation&&"idle"===e.revalidation?{error:e.error,location:e.location,revalidation:e.revalidation}:{error:void 0!==e.error?e.error:t.error,location:t.location,revalidation:e.revalidation||t.revalidation}}componentDidCatch(e,t){console.error("React Router caught the following error during render",e,t)}render(){return void 0!==this.state.error?r.createElement(J.Provider,{value:this.props.routeContext},r.createElement(K.Provider,{value:this.state.error,children:this.props.component})):this.props.children}};function ne({routeContext:e,match:t,children:a}){let n=r.useContext(B);return n&&n.static&&n.staticContext&&(t.route.errorElement||t.route.ErrorBoundary)&&(n.staticContext._deepestRenderedBoundaryId=t.route.id),r.createElement(J.Provider,{value:e},a)}function re(e){return`${e} must be used within a data router.  See https://reactrouter.com/en/main/routers/picking-a-router.`}function le(e){let t=function(e){let t=r.useContext(J);return c(t,re(e)),t}(e),a=t.matches[t.matches.length-1];return c(a.route.id,`${e} can only be used on routes that contain a unique "id"`),a.route.id}var oe={};function ie(e,t,a){t||oe[e]||(oe[e]=!0,u(!1,a))}function se(e){c(!1,"A <Route> is only ever to be used as the child of <Routes> element, never rendered directly. Please wrap your <Route> in a <Routes>.")}function ce({basename:e="/",children:t=null,location:a,navigationType:n="POP",navigator:l,static:o=!1}){c(!z(),"You cannot render a <Router> inside another <Router>. You should never have more than one in your app.");let i=e.replace(/^\/*/,"/"),s=r.useMemo((()=>({basename:i,navigator:l,static:o,future:{}})),[i,l,o]);"string"==typeof a&&(a=h(a));let{pathname:m="/",search:d="",hash:p="",state:f=null,key:y="default"}=a,g=r.useMemo((()=>{let e=$(m,i);return null==e?null:{location:{pathname:e,search:d,hash:p,state:f,key:y},navigationType:n}}),[i,m,d,p,f,y,n]);return u(null!=g,`<Router basename="${i}"> is not able to match the URL "${m}${d}${p}" because it does not start with the basename, so the <Router> won't render anything.`),null==g?null:r.createElement(j.Provider,{value:s},r.createElement(H.Provider,{children:t,value:g}))}function ue({children:e,location:t}){return Z(me(e),t)}function me(e,t=[]){let a=[];return r.Children.forEach(e,((e,n)=>{if(!r.isValidElement(e))return;let l=[...t,n];if(e.type===r.Fragment)return void a.push.apply(a,me(e.props.children,l));c(e.type===se,`[${"string"==typeof e.type?e.type:e.type.name}] is not a <Route> component. All component children of <Routes> must be a <Route> or <React.Fragment>`),c(!e.props.index||!e.props.children,"An index route cannot have child routes.");let o={id:e.props.id||l.join("-"),caseSensitive:e.props.caseSensitive,element:e.props.element,Component:e.props.Component,index:e.props.index,path:e.props.path,loader:e.props.loader,action:e.props.action,hydrateFallbackElement:e.props.hydrateFallbackElement,HydrateFallback:e.props.HydrateFallback,errorElement:e.props.errorElement,ErrorBoundary:e.props.ErrorBoundary,hasErrorBoundary:!0===e.props.hasErrorBoundary||null!=e.props.ErrorBoundary||null!=e.props.errorElement,shouldRevalidate:e.props.shouldRevalidate,handle:e.props.handle,lazy:e.props.lazy};e.props.children&&(o.children=me(e.props.children,l)),a.push(o)})),a}r.memo((function({routes:e,future:t,state:a}){return Z(e,void 0,a,t)})),r.Component;var de="get",pe="application/x-www-form-urlencoded";function he(e){return null!=e&&"string"==typeof e.tagName}var fe=null,ye=new Set(["application/x-www-form-urlencoded","multipart/form-data","text/plain"]);function ge(e){return null==e||ye.has(e)?e:(u(!1,`"${e}" is not a valid \`encType\` for \`<Form>\`/\`<fetcher.Form>\` and will default to "${pe}"`),null)}function Ee(e,t){if(!1===e||null==e)throw new Error(t)}function ve(e){return null!=e&&(null==e.href?"preload"===e.rel&&"string"==typeof e.imageSrcSet&&"string"==typeof e.imageSizes:"string"==typeof e.rel&&"string"==typeof e.href)}function we(e,t,a,n,r,l){let o=(e,t)=>!a[t]||e.route.id!==a[t].route.id,i=(e,t)=>a[t].pathname!==e.pathname||a[t].route.path?.endsWith("*")&&a[t].params["*"]!==e.params["*"];return"assets"===l?t.filter(((e,t)=>o(e,t)||i(e,t))):"data"===l?t.filter(((t,l)=>{let s=n.routes[t.route.id];if(!s||!s.hasLoader)return!1;if(o(t,l)||i(t,l))return!0;if(t.route.shouldRevalidate){let n=t.route.shouldRevalidate({currentUrl:new URL(r.pathname+r.search+r.hash,window.origin),currentParams:a[0]?.params||{},nextUrl:new URL(e,window.origin),nextParams:t.params,defaultShouldRevalidate:!0});if("boolean"==typeof n)return n}return!0})):[]}function be(e,t,{includeHydrateFallback:a}={}){return n=e.map((e=>{let n=t.routes[e.route.id];if(!n)return[];let r=[n.module];return n.clientActionModule&&(r=r.concat(n.clientActionModule)),n.clientLoaderModule&&(r=r.concat(n.clientLoaderModule)),a&&n.hydrateFallbackModule&&(r=r.concat(n.hydrateFallbackModule)),n.imports&&(r=r.concat(n.imports)),r})).flat(1),[...new Set(n)];var n}Object.getOwnPropertyNames(Object.prototype).sort().join("\0"),"undefined"!=typeof window?window:"undefined"!=typeof globalThis&&globalThis,Symbol("SingleFetchRedirect");function Ne(){let e=r.useContext(B);return Ee(e,"You must render this element inside a <DataRouterContext.Provider> element"),e}function _e(){let e=r.useContext(I);return Ee(e,"You must render this element inside a <DataRouterStateContext.Provider> element"),e}r.Component;var Se=r.createContext(void 0);function Ce(){let e=r.useContext(Se);return Ee(e,"You must render this element inside a <HydratedRouter> element"),e}function xe(e,t){return a=>{e&&e(a),a.defaultPrevented||t(a)}}function ke({page:e,...t}){let{router:a}=Ne(),n=r.useMemo((()=>f(a.routes,e,a.basename)),[a.routes,e,a.basename]);return n?r.createElement($e,{page:e,matches:n,...t}):null}function Re(e){let{manifest:t,routeModules:a}=Ce(),[n,l]=r.useState([]);return r.useEffect((()=>{let n=!1;return async function(e,t,a){return function(e,t){let a=new Set,n=new Set(t);return e.reduce(((e,r)=>{if(t&&(null==(l=r)||"string"!=typeof l.page)&&"script"===r.as&&r.href&&n.has(r.href))return e;var l;let o=JSON.stringify(function(e){let t={},a=Object.keys(e).sort();for(let n of a)t[n]=e[n];return t}(r));return a.has(o)||(a.add(o),e.push({key:o,link:r})),e}),[])}((await Promise.all(e.map((async e=>{let n=t.routes[e.route.id];if(n){let e=await async function(e,t){if(e.id in t)return t[e.id];try{let a=await import(e.module);return t[e.id]=a,a}catch(t){return console.error(`Error loading route module \`${e.module}\`, reloading page...`),console.error(t),window.__reactRouterContext&&window.__reactRouterContext.isSpaMode,window.location.reload(),new Promise((()=>{}))}}(n,a);return e.links?e.links():[]}return[]})))).flat(1).filter(ve).filter((e=>"stylesheet"===e.rel||"preload"===e.rel)).map((e=>"stylesheet"===e.rel?{...e,rel:"prefetch",as:"style"}:{...e,rel:"prefetch"})))}(e,t,a).then((e=>{n||l(e)})),()=>{n=!0}}),[e,t,a]),n}function $e({page:e,matches:t,...a}){let n=Y(),{manifest:l,routeModules:o}=Ce(),{basename:i}=Ne(),{loaderData:s,matches:c}=_e(),u=r.useMemo((()=>we(e,t,c,l,n,"data")),[e,t,c,l,n]),m=r.useMemo((()=>we(e,t,c,l,n,"assets")),[e,t,c,l,n]),d=r.useMemo((()=>{if(e===n.pathname+n.search+n.hash)return[];let a=new Set,r=!1;if(t.forEach((e=>{let t=l.routes[e.route.id];t&&t.hasLoader&&(!u.some((t=>t.route.id===e.route.id))&&e.route.id in s&&o[e.route.id]?.shouldRevalidate||t.hasClientLoader?r=!0:a.add(e.route.id))})),0===a.size)return[];let c=function(e,t){let a="string"==typeof e?new URL(e,"undefined"==typeof window?"server://singlefetch/":window.location.origin):e;return"/"===a.pathname?a.pathname="_root.data":t&&"/"===$(a.pathname,t)?a.pathname=`${t.replace(/\/$/,"")}/_root.data`:a.pathname=`${a.pathname.replace(/\/$/,"")}.data`,a}(e,i);return r&&a.size>0&&c.searchParams.set("_routes",t.filter((e=>a.has(e.route.id))).map((e=>e.route.id)).join(",")),[c.pathname+c.search]}),[i,s,n,l,u,t,e,o]),p=r.useMemo((()=>be(m,l)),[m,l]),h=Re(m);return r.createElement(r.Fragment,null,d.map((e=>r.createElement("link",{key:e,rel:"prefetch",as:"fetch",href:e,...a}))),p.map((e=>r.createElement("link",{key:e,rel:"modulepreload",href:e,...a}))),h.map((({key:e,link:t})=>r.createElement("link",{key:e,...t}))))}Se.displayName="FrameworkContext";function Pe(...e){return t=>{e.forEach((e=>{"function"==typeof e?e(t):null!=e&&(e.current=t)}))}}var Le="undefined"!=typeof window&&void 0!==window.document&&void 0!==window.document.createElement;try{Le&&(window.__reactRouterVersion="7.6.1")}catch(e){}var Te=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,Fe=r.forwardRef((function({onClick:e,discover:t="render",prefetch:a="none",relative:n,reloadDocument:l,replace:o,state:i,target:s,to:m,preventScrollReset:d,viewTransition:h,...f},y){let g,{basename:E}=r.useContext(j),v="string"==typeof m&&Te.test(m),w=!1;if("string"==typeof m&&v&&(g=m,Le))try{let e=new URL(window.location.href),t=m.startsWith("//")?new URL(e.protocol+m):new URL(m),a=$(t.pathname,E);t.origin===e.origin&&null!=a?m=a+t.search+t.hash:w=!0}catch(e){u(!1,`<Link to="${m}"> contains an invalid URL which will probably break when clicked - please update to a valid URL path.`)}let b=function(e,{relative:t}={}){c(z(),"useHref() may be used only in the context of a <Router> component.");let{basename:a,navigator:n}=r.useContext(j),{hash:l,pathname:o,search:i}=X(e,{relative:t}),s=o;return"/"!==a&&(s="/"===o?a:F([a,o])),n.createHref({pathname:s,search:i,hash:l})}(m,{relative:n}),[N,_,S]=function(e,t){let a=r.useContext(Se),[n,l]=r.useState(!1),[o,i]=r.useState(!1),{onFocus:s,onBlur:c,onMouseEnter:u,onMouseLeave:m,onTouchStart:d}=t,p=r.useRef(null);r.useEffect((()=>{if("render"===e&&i(!0),"viewport"===e){let e=new IntersectionObserver((e=>{e.forEach((e=>{i(e.isIntersecting)}))}),{threshold:.5});return p.current&&e.observe(p.current),()=>{e.disconnect()}}}),[e]),r.useEffect((()=>{if(n){let e=setTimeout((()=>{i(!0)}),100);return()=>{clearTimeout(e)}}}),[n]);let h=()=>{l(!0)},f=()=>{l(!1),i(!1)};return a?"intent"!==e?[o,p,{}]:[o,p,{onFocus:xe(s,h),onBlur:xe(c,f),onMouseEnter:xe(u,h),onMouseLeave:xe(m,f),onTouchStart:xe(d,h)}]:[!1,p,{}]}(a,f),C=function(e,{target:t,replace:a,state:n,preventScrollReset:l,relative:o,viewTransition:i}={}){let s=G(),c=Y(),u=X(e,{relative:o});return r.useCallback((r=>{if(function(e,t){return!(0!==e.button||t&&"_self"!==t||function(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}(e))}(r,t)){r.preventDefault();let t=void 0!==a?a:p(c)===p(u);s(e,{replace:t,state:n,preventScrollReset:l,relative:o,viewTransition:i})}}),[c,s,u,a,n,t,e,l,o,i])}(m,{replace:o,state:i,target:s,preventScrollReset:d,relative:n,viewTransition:h}),x=r.createElement("a",{...f,...S,href:g||b,onClick:w||l?e:function(t){e&&e(t),t.defaultPrevented||C(t)},ref:Pe(y,_),target:s,"data-discover":v||"render"!==t?void 0:"true"});return N&&!v?r.createElement(r.Fragment,null,x,r.createElement(ke,{page:b})):x}));function Oe(e){let t=r.useContext(B);return c(t,function(e){return`${e} must be used within a data router.  See https://reactrouter.com/en/main/routers/picking-a-router.`}(e)),t}Fe.displayName="Link",r.forwardRef((function({"aria-current":e="page",caseSensitive:t=!1,className:a="",end:n=!1,style:l,to:o,viewTransition:i,children:s,...u},m){let d=X(o,{relative:u.relative}),p=Y(),h=r.useContext(I),{navigator:f,basename:y}=r.useContext(j),g=null!=h&&function(e,t={}){let a=r.useContext(U);c(null!=a,"`useViewTransitionState` must be used within `react-router-dom`'s `RouterProvider`.  Did you accidentally import `RouterProvider` from `react-router`?");let{basename:n}=Oe("useViewTransitionState"),l=X(e,{relative:t.relative});if(!a.isTransitioning)return!1;let o=$(a.currentLocation.pathname,n)||a.currentLocation.pathname,i=$(a.nextLocation.pathname,n)||a.nextLocation.pathname;return null!=k(l.pathname,i)||null!=k(l.pathname,o)}(d)&&!0===i,E=f.encodeLocation?f.encodeLocation(d).pathname:d.pathname,v=p.pathname,w=h&&h.navigation&&h.navigation.location?h.navigation.location.pathname:null;t||(v=v.toLowerCase(),w=w?w.toLowerCase():null,E=E.toLowerCase()),w&&y&&(w=$(w,y)||w);const b="/"!==E&&E.endsWith("/")?E.length-1:E.length;let N,_=v===E||!n&&v.startsWith(E)&&"/"===v.charAt(b),S=null!=w&&(w===E||!n&&w.startsWith(E)&&"/"===w.charAt(E.length)),C={isActive:_,isPending:S,isTransitioning:g},x=_?e:void 0;N="function"==typeof a?a(C):[a,_?"active":null,S?"pending":null,g?"transitioning":null].filter(Boolean).join(" ");let R="function"==typeof l?l(C):l;return r.createElement(Fe,{...u,"aria-current":x,className:N,ref:m,style:R,to:o,viewTransition:i},"function"==typeof s?s(C):s)})).displayName="NavLink",r.forwardRef((({discover:e="render",fetcherKey:t,navigate:a,reloadDocument:n,replace:l,state:o,method:i=de,action:s,onSubmit:u,relative:m,preventScrollReset:d,viewTransition:h,...f},y)=>{let g=function(){let{router:e}=Oe("useSubmit"),{basename:t}=r.useContext(j),a=le("useRouteId");return r.useCallback((async(n,r={})=>{let{action:l,method:o,encType:i,formData:s,body:c}=function(e,t){let a,n,r,l,o;if(he(i=e)&&"form"===i.tagName.toLowerCase()){let o=e.getAttribute("action");n=o?$(o,t):null,a=e.getAttribute("method")||de,r=ge(e.getAttribute("enctype"))||pe,l=new FormData(e)}else if(function(e){return he(e)&&"button"===e.tagName.toLowerCase()}(e)||function(e){return he(e)&&"input"===e.tagName.toLowerCase()}(e)&&("submit"===e.type||"image"===e.type)){let o=e.form;if(null==o)throw new Error('Cannot submit a <button> or <input type="submit"> without a <form>');let i=e.getAttribute("formaction")||o.getAttribute("action");if(n=i?$(i,t):null,a=e.getAttribute("formmethod")||o.getAttribute("method")||de,r=ge(e.getAttribute("formenctype"))||ge(o.getAttribute("enctype"))||pe,l=new FormData(o,e),!function(){if(null===fe)try{new FormData(document.createElement("form"),0),fe=!1}catch(e){fe=!0}return fe}()){let{name:t,type:a,value:n}=e;if("image"===a){let e=t?`${t}.`:"";l.append(`${e}x`,"0"),l.append(`${e}y`,"0")}else t&&l.append(t,n)}}else{if(he(e))throw new Error('Cannot submit element that is not <form>, <button>, or <input type="submit|image">');a=de,n=null,r=pe,o=e}var i;return l&&"text/plain"===r&&(o=l,l=void 0),{action:n,method:a.toLowerCase(),encType:r,formData:l,body:o}}(n,t);if(!1===r.navigate){let t=r.fetcherKey||Me();await e.fetch(t,a,r.action||l,{preventScrollReset:r.preventScrollReset,formData:s,body:c,formMethod:r.method||o,formEncType:r.encType||i,flushSync:r.flushSync})}else await e.navigate(r.action||l,{preventScrollReset:r.preventScrollReset,formData:s,body:c,formMethod:r.method||o,formEncType:r.encType||i,replace:r.replace,state:r.state,fromRouteId:a,flushSync:r.flushSync,viewTransition:r.viewTransition})}),[e,t,a])}(),E=function(e,{relative:t}={}){let{basename:a}=r.useContext(j),n=r.useContext(J);c(n,"useFormAction must be used inside a RouteContext");let[l]=n.matches.slice(-1),o={...X(e||".",{relative:t})},i=Y();if(null==e){o.search=i.search;let e=new URLSearchParams(o.search),t=e.getAll("index");if(t.some((e=>""===e))){e.delete("index"),t.filter((e=>e)).forEach((t=>e.append("index",t)));let a=e.toString();o.search=a?`?${a}`:""}}return e&&"."!==e||!l.route.index||(o.search=o.search?o.search.replace(/^\?/,"?index&"):"?index"),"/"!==a&&(o.pathname="/"===o.pathname?a:F([a,o.pathname])),p(o)}(s,{relative:m}),v="get"===i.toLowerCase()?"get":"post",w="string"==typeof s&&Te.test(s);return r.createElement("form",{ref:y,method:v,action:E,onSubmit:n?u:e=>{if(u&&u(e),e.defaultPrevented)return;e.preventDefault();let n=e.nativeEvent.submitter,r=n?.getAttribute("formmethod")||i;g(n||e.currentTarget,{fetcherKey:t,method:r,navigate:a,replace:l,state:o,relative:m,preventScrollReset:d,viewTransition:h})},...f,"data-discover":w||"render"!==e?void 0:"true"})})).displayName="Form";var De=0,Me=()=>`__${String(++De)}__`;const Ae=window.wp.apiFetch;var We=a.n(Ae);const qe=()=>{const[e,t]=(0,r.useState)(!1),[a,o]=(0,r.useState)(!1),[i,s]=(0,r.useState)("grid"),[c,u]=(0,r.useState)(!1),[m,d]=(0,r.useState)(!1),[p,h]=(0,r.useState)(!1),[f,y]=(0,r.useState)([]),[g,E]=(0,r.useState)(1),[v,w]=(0,r.useState)(1),[b,N]=(0,r.useState)(!1),[_,S]=(0,r.useState)({}),[C,x]=(0,r.useState)(!0),[k,R]=(0,r.useState)({}),[$,P]=(0,r.useState)([]),[L,T]=(0,r.useState)(0),[F,O]=(0,r.useState)(""),[D,M]=(0,r.useState)([]),[A,W]=(0,r.useState)({}),[q,B]=(0,r.useState)({gateway:{id:"",title:""}}),[I,U]=(0,r.useState)({}),[j,H]=(0,r.useState)(!1),[J,K]=(0,r.useState)(null),[z,Y]=(0,r.useState)(""),[Q,V]=(0,r.useState)([]),[G,X]=(0,r.useState)({}),[Z,ee]=(0,r.useState)({line_items:[],fee_lines:[],coupon_lines:[]}),[te,ae]=(0,r.useState)({customer_id:0,customer_note:"",payment_method:"",payment_method_title:"",billing:{},shipping:{}}),ne=(0,r.useRef)(null),re=(0,r.useRef)(null),le=(0,r.useCallback)((()=>{const e=new URLSearchParams(window.location.search).get("category");return null!==e?f.filter((t=>void 0!==t.categories.find((t=>t.id===parseInt(e))))):f}),[f]),oe=(0,r.useCallback)((()=>{const e=(parseFloat(F.replace(/[^\d.-]/g,""))||0)-(de()||0);return e>0?e:0}),[F,Z,A]),ie=(0,r.useCallback)((()=>{let e=Z.line_items.length>0&&""!==z;return"wepos_cash"===z&&(e=(parseFloat(F.replace(/[^\d.-]/g,""))||0)>=(de()||0)&&e),e}),[Z.line_items,z,F,A]),se=e=>{const t="number"==typeof e?e:parseFloat(String(e||0));return isNaN(t)?"$0.00":`$${t.toFixed(2)}`},ce=e=>!e.manage_stock||"instock"===e.stock_status&&e.stock_quantity>0,ue=()=>Z.line_items.reduce(((e,t)=>e+(t.on_sale?t.sale_price||0:t.regular_price||0)*(t.quantity||0)),0),me=()=>Z.line_items.reduce(((e,t)=>e+(t.total_tax||0)),0),de=()=>{let e=ue()||0;return e+=Z.fee_lines.reduce(((e,t)=>e+(t.total||0)),0),e-=Z.coupon_lines.reduce(((e,t)=>"discount"===t.type?e+Math.abs(t.total||0):e),0),"incl"!==A.woo_tax?.wc_tax_display_cart&&(e+=me()),e},pe=e=>e.images.length>0?e.images[0].woocommerce_thumbnail:window.wepos?.placeholder_image||"",he=async()=>{if(1===v&&u(!0),g>=v)try{const e=await We()({path:`/${window.wepos.rest.posversion}/products?status=publish&per_page=30&page=${v}`});fe(e),w((e=>e+1)),u(!1),30===e.length&&setTimeout(he,100)}catch(e){console.error("Error fetching products:",e),u(!1)}else u(!1)},fe=e=>{const t=e.filter((e=>"variable"!==e.type||!ye(e)));y((e=>[...e,...t]))},ye=e=>!e.attributes||!e.attributes.some((e=>!0===e.variation)),ge=(e,t)=>{const a=[...Z.line_items];a[t].editQuantity=!a[t].editQuantity,ee((e=>({...e,line_items:a})))},Ee=()=>{ee({line_items:[],fee_lines:[],coupon_lines:[]}),ae({customer_id:0,customer_note:"",payment_method:"",payment_method_title:"",billing:{},shipping:{}}),h(!1),O(""),o(!1)},ve=()=>{Ee(),window.history.pushState({},"",window.location.pathname)},we=()=>{Z.line_items.length<=0||(d(!0),$.length>0&&(Y($[0].id),ae((e=>({...e,payment_method:$[0].id,payment_method_title:$[0].title})))))},be=()=>{d(!1),t(!1)};return(0,r.useEffect)((()=>{const e=e=>{switch(e.key){case"F3":e.preventDefault(),s((e=>"grid"===e?"list":"grid"));break;case"F8":e.preventDefault(),e.shiftKey?Ee():ve();break;case"F9":e.preventDefault(),we();break;case"Escape":e.preventDefault(),be();break;case"/":(e.ctrlKey||e.metaKey)&&(e.preventDefault(),t((e=>!e)))}};return window.addEventListener("keydown",e),()=>window.removeEventListener("keydown",e)}),[]),(0,r.useEffect)((()=>{if((async()=>{try{const e=await We()({path:`/${window.wepos.rest.posversion}/settings`});W(e)}catch(e){console.error("Error fetching settings:",e)}})(),he(),(async()=>{try{const e=await We()({path:`/${window.wepos.rest.posversion}/payment/gateways`});P(e),T(4-e.length%4)}catch(e){console.error("Error fetching gateways:",e)}})(),(async()=>{try{const e=await We()({path:`/${window.wepos.rest.wcversion}/products/categories?hide_empty=true&_fields=id,name,parent_id&per_page=100`});e.sort(((e,t)=>e.name.localeCompare(t.name)));const t=(e,a=null,n=0)=>e.filter((e=>e.parent_id===a)).map((a=>[{...a,level:n},...t(e,a.id,n+1)])).flat(),a=t(e),n={id:-1,level:0,name:"All categories",parent_id:null};V([n,...a]),K(n)}catch(e){console.error("Error fetching categories:",e)}})(),"undefined"!=typeof localStorage)try{const e=localStorage.getItem("cartdata"),t=localStorage.getItem("orderdata");e&&ee(JSON.parse(e)),t&&ae(JSON.parse(t))}catch(e){console.error("Error loading from localStorage:",e)}const e=()=>{"undefined"!=typeof localStorage&&(localStorage.setItem("cartdata",JSON.stringify(Z)),localStorage.setItem("orderdata",JSON.stringify(te)))};return window.addEventListener("beforeunload",e),()=>window.removeEventListener("beforeunload",e)}),[]),(0,r.useEffect)((()=>{p&&j&&(setTimeout((()=>{window.confirm("Order successful! Would you like to print the receipt?")&&window.print()}),500),H(!1))}),[p,j]),(0,r.useEffect)((()=>{"undefined"!=typeof localStorage&&localStorage.setItem("cartdata",JSON.stringify(Z))}),[Z]),(0,r.useEffect)((()=>{"undefined"!=typeof localStorage&&localStorage.setItem("orderdata",JSON.stringify(te))}),[te]),(0,n.createElement)("div",{id:"wepos-main"},(0,n.createElement)("div",{className:"content-product"},(0,n.createElement)("div",{className:"top-panel wepos-clearfix"},(0,n.createElement)("div",{className:"search-bar"},(0,n.createElement)("div",{className:"search-box"},(0,n.createElement)("input",{type:"text",id:"product-search",placeholder:"Search Product By Name, Barcode, SKU..."}),(0,n.createElement)("span",{className:"search-icon flaticon-search"}))),(0,n.createElement)("div",{className:"category"},(0,n.createElement)("select",{id:"product-category"},Q.map((e=>(0,n.createElement)("option",{key:e.id,value:e.id},"  ".repeat(e.level),e.name))))),(0,n.createElement)("div",{className:"toggle-view"},(0,n.createElement)("div",{className:"product-toggle"},(0,n.createElement)("span",{className:"toggle-icon list-view flaticon-menu-button-of-three-horizontal-lines "+("list"===i?"active":""),onClick:()=>s("list")}),(0,n.createElement)("span",{className:"toggle-icon grid-view flaticon-menu "+("grid"===i?"active":""),onClick:()=>s("grid")})))),(0,n.createElement)("div",{className:`items-wrapper ${i}`,ref:ne},c?(0,n.createElement)("div",{className:"product-loading"},(0,n.createElement)("div",{className:"spinner spinner-loading"})):(0,n.createElement)(n.Fragment,null,le().length>0?le().map((e=>{return(0,n.createElement)("div",{key:e.id,className:"item"},"simple"===e.type&&(0,n.createElement)("div",{className:"item-wrap "+(ce(e)?"":"disabled"),onClick:()=>(e=>{if(!ce(e))return void alert("Product is out of stock!");const t=Z.line_items.findIndex((t=>t.product_id===e.id));if(-1!==t){const e=[...Z.line_items];e[t].quantity+=1,ee((t=>({...t,line_items:e})))}else{const t={id:Date.now(),product_id:e.id,name:e.name,quantity:1,type:e.type,on_sale:e.on_sale,sale_price:e.sale_price,regular_price:e.regular_price,attribute:[]};ee((e=>({...e,line_items:[...e.line_items,t]})))}})(e)},(0,n.createElement)("div",{className:"img"},(0,n.createElement)("img",{src:pe(e),alt:e.name})),(0,n.createElement)("div",{className:"title"},"grid"===i?(t=e.name).length>20?t.substring(0,20)+"...":t:(0,n.createElement)("div",null,(0,n.createElement)("div",{className:"product-name"},e.name),(0,n.createElement)("ul",{className:"meta"},e.sku&&(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"label"},"SKU:"),(0,n.createElement)("span",{className:"value"},e.sku)),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"label"},"Price:"),(0,n.createElement)("span",{className:"value",dangerouslySetInnerHTML:{__html:e.price_html}}))))),(0,n.createElement)("span",{className:`add-product-icon flaticon-add ${i}`})));var t})):(0,n.createElement)("div",{className:"no-product-found"},(0,n.createElement)("img",{src:`${window.wepos?.assets_url}/images/no-product.png`,alt:"",width:"120px"}),(0,n.createElement)("p",null,"No Product Found"))))),(0,n.createElement)("div",{className:"content-cart"},(0,n.createElement)("div",{className:"top-panel"},(0,n.createElement)("div",{className:"customer-search-box"},(0,n.createElement)("input",{type:"text",id:"customer-search",placeholder:"Walk-in Customer"}),(0,n.createElement)("span",{className:"add-new-customer flaticon-add"})),(0,n.createElement)("div",{className:"action"},(0,n.createElement)("div",{className:"more-options"},(0,n.createElement)("button",{className:"wepos-button",onClick:()=>o(!a)},(0,n.createElement)("span",{className:"more-icon flaticon-more"})),a&&(0,n.createElement)("div",{className:"wepos-dropdown-menu"},(0,n.createElement)("ul",null,(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:Ee},"Empty Cart")),(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:()=>t(!0)},"Help")),(0,n.createElement)("li",{className:"divider"}),(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:()=>window.location.href=window.wepos?.logout_url},"Logout"))))))),A.wepos_general&&(0,n.createElement)("div",{className:"cart-panel"},(0,n.createElement)("div",{className:"cart-content"},(0,n.createElement)("table",{className:"cart-table"},(0,n.createElement)("thead",null,(0,n.createElement)("tr",null,(0,n.createElement)("th",{style:{width:"65%"}},"Product"),(0,n.createElement)("th",{style:{width:"15%"}},"Qty"),(0,n.createElement)("th",{style:{width:"30%"}},"Price"),(0,n.createElement)("th",null),(0,n.createElement)("th",null))),(0,n.createElement)("tbody",null,Z.line_items.length>0?Z.line_items.map(((e,t)=>(0,n.createElement)(l().Fragment,{key:e.id},(0,n.createElement)("tr",null,(0,n.createElement)("td",{className:"name",onClick:()=>ge(0,t)},e.name),(0,n.createElement)("td",{className:"qty",onClick:()=>ge(0,t)},e.quantity),(0,n.createElement)("td",{className:"price",onClick:()=>ge(0,t)},e.on_sale?(0,n.createElement)(n.Fragment,null,(0,n.createElement)("span",{className:"sale-price"},se(e.quantity*e.sale_price)),(0,n.createElement)("span",{className:"regular-price"},se(e.quantity*e.regular_price))):(0,n.createElement)("span",{className:"sale-price"},se(e.quantity*e.regular_price))),(0,n.createElement)("td",{className:"action"},(0,n.createElement)("span",{className:"flaticon-right-arrow "+(e.editQuantity?"open":""),onClick:()=>ge(0,t)})),(0,n.createElement)("td",{className:"remove"},(0,n.createElement)("span",{className:"flaticon-cancel-music",onClick:()=>(e=>{const t=Z.line_items.filter(((t,a)=>a!==e));ee((e=>({...e,line_items:t})))})(t)}))),e.editQuantity&&(0,n.createElement)("tr",{className:"update-quantity-wrap"},(0,n.createElement)("td",{colSpan:5},(0,n.createElement)("span",{className:"qty"},"Quantity"),(0,n.createElement)("span",{className:"qty-number"},(0,n.createElement)("input",{type:"number",min:"1",step:"1",value:e.quantity,onChange:e=>{const a=[...Z.line_items];a[t].quantity=parseInt(e.target.value)||1,ee((e=>({...e,line_items:a})))}})),(0,n.createElement)("span",{className:"qty-action"},(0,n.createElement)("a",{href:"#",className:"add",onClick:e=>{e.preventDefault(),((e,t)=>{const a=[...Z.line_items];a[t].quantity+=1,ee((e=>({...e,line_items:a})))})(0,t)}},"+"),(0,n.createElement)("a",{href:"#",className:"minus",onClick:e=>{e.preventDefault(),((e,t)=>{const a=[...Z.line_items];a[t].quantity>1&&(a[t].quantity-=1,ee((e=>({...e,line_items:a}))))})(0,t)}},"-"))))))):(0,n.createElement)("tr",{className:"no-item"},(0,n.createElement)("td",{colSpan:5},(0,n.createElement)("img",{src:`${window.wepos?.assets_url}/images/empty-cart.png`,alt:"",width:"120px"}),(0,n.createElement)("p",null,"Empty Cart")))))),(0,n.createElement)("div",{className:"cart-calculation"},(0,n.createElement)("table",{className:"cart-total-table"},(0,n.createElement)("tbody",null,(0,n.createElement)("tr",{className:"cart-meta-data"},(0,n.createElement)("td",{className:"label"},"Subtotal","incl"===A.woo_tax?.wc_tax_display_cart&&me()>0&&(0,n.createElement)("span",{className:"name"},"Including Tax")),(0,n.createElement)("td",{className:"price"},se(ue())),(0,n.createElement)("td",{className:"action"})),me()>0&&(0,n.createElement)("tr",{className:"tax"},(0,n.createElement)("td",{className:"label"},"incl"===A.woo_tax?.wc_tax_display_cart?"Fee Tax":"Tax"),(0,n.createElement)("td",{className:"price"},se(me())),(0,n.createElement)("td",{className:"action"})),(0,n.createElement)("tr",{className:"pay-now",onClick:we},(0,n.createElement)("td",null,"Pay Now"),(0,n.createElement)("td",{className:"amount"},se(de())),(0,n.createElement)("td",{className:"icon"},(0,n.createElement)("span",{className:"flaticon-right-arrow"})))))))),e&&(0,n.createElement)("div",{className:"modal-overlay",onClick:()=>t(!1)},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-help-wrapper"},(0,n.createElement)("h2",null,"Shortcut Keys"),(0,n.createElement)("ul",null,(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f3")),(0,n.createElement)("span",{className:"title"},"Toggle Product View")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f8")),(0,n.createElement)("span",{className:"title"},"Create New Sale")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"shift+f8")),(0,n.createElement)("span",{className:"title"},"Empty your cart")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f9")),(0,n.createElement)("span",{className:"title"},"Process Payment")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"ctrl/cmd+?")),(0,n.createElement)("span",{className:"title"},"Show/Close Help")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"esc")),(0,n.createElement)("span",{className:"title"},"Close anything")))))),m&&(0,n.createElement)("div",{className:"modal-overlay",onClick:be},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-checkout-wrapper"},(0,n.createElement)("div",{className:"left-content"},(0,n.createElement)("div",{className:"header"},"Sale Summary"),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("table",{className:"sale-summary-cart"},(0,n.createElement)("tbody",null,Z.line_items.map((e=>(0,n.createElement)("tr",{key:e.id},(0,n.createElement)("td",{className:"name"},e.name),(0,n.createElement)("td",{className:"quantity"},e.quantity),(0,n.createElement)("td",{className:"price"},e.on_sale?(0,n.createElement)(n.Fragment,null,(0,n.createElement)("span",{className:"sale-price"},se(e.quantity*e.sale_price)),(0,n.createElement)("span",{className:"regular-price"},se(e.quantity*e.regular_price))):(0,n.createElement)("span",{className:"sale-price"},se(e.quantity*e.regular_price)))))))))),(0,n.createElement)("div",{className:"right-content"},(0,n.createElement)("div",{className:"header"},(0,n.createElement)("h2",null,"Pay"),(0,n.createElement)("span",{className:"pay-amount"},se(de()))),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("div",{className:"payment-gateway"},$.length>0?$.map((e=>(0,n.createElement)("label",{key:e.id},(0,n.createElement)("input",{type:"radio",name:"gateway",value:e.id,checked:z===e.id,onChange:e=>Y(e.target.value)}),(0,n.createElement)("span",{className:`gateway gateway-${e.id}`},e.title)))):(0,n.createElement)("p",null,"No gateway found")),"wepos_cash"===z&&(0,n.createElement)("div",{className:"payment-option"},(0,n.createElement)("div",{className:"payment-amount"},(0,n.createElement)("div",{className:"input-part"},(0,n.createElement)("div",{className:"input-wrap"},(0,n.createElement)("p",null,"Cash"),(0,n.createElement)("div",{className:"input-addon"},(0,n.createElement)("span",{className:"currency"},"$"),(0,n.createElement)("input",{id:"input-cash-amount",type:"text",value:F,onChange:e=>O(e.target.value),ref:re})))),(0,n.createElement)("div",{className:"change-money"},(0,n.createElement)("p",null,"Change money: ",se(oe())))))),(0,n.createElement)("div",{className:"footer"},(0,n.createElement)("a",{href:"#",className:"back-btn",onClick:e=>{e.preventDefault(),be()}},"Back to Sale"),(0,n.createElement)("button",{className:"process-checkout-btn "+(ie()?"":"disabled"),onClick:async()=>{if(ie())try{const e=document.querySelector(".wepos-checkout-wrapper");e&&(e.style.opacity="0.6",e.style.pointerEvents="none");const t={billing:te.billing,shipping:te.shipping,line_items:Z.line_items.map((e=>({product_id:e.product_id,quantity:e.quantity}))),fee_lines:Z.fee_lines,coupon_lines:Z.coupon_lines,customer_id:te.customer_id,customer_note:te.customer_note,payment_method:z,payment_method_title:$.find((e=>e.id===z))?.title||"",meta_data:[{key:"_wepos_is_pos_order",value:!0},{key:"_wepos_cash_tendered_amount",value:F.toString()},{key:"_wepos_cash_change_amount",value:oe().toString()}]};console.log("Sending order payload:",t);const a=await We()({path:`/${window.wepos.rest.wcversion}/orders`,method:"POST",data:t}),n={};a.line_items.forEach((e=>{n[e.product_id]=e.total_tax}));const r=Z.line_items.map((e=>({...e,total_tax:n[e.product_id]||0})));ee((e=>({...e,line_items:r}))),"success"===(await We()({path:`/${window.wepos.rest.posversion}/payment/process`,method:"POST",data:a})).result&&(B({line_items:Z.line_items,fee_lines:Z.fee_lines,coupon_lines:Z.coupon_lines,subtotal:ue(),taxtotal:me(),ordertotal:de(),gateway:{id:a.payment_method,title:a.payment_method_title},order_id:a.number,order_date:a.date_created,cashamount:F.toString(),changeamount:oe().toString()}),d(!1),h(!0),H(!0)),e&&(e.style.opacity="1",e.style.pointerEvents="auto")}catch(e){const t=document.querySelector(".wepos-checkout-wrapper");t&&(t.style.opacity="1",t.style.pointerEvents="auto"),alert(e?.message||"Payment processing failed"),console.error("Payment processing error:",e)}},disabled:!ie()},"Process Payment")))))),p&&(0,n.createElement)("div",{className:"modal-overlay",onClick:()=>h(!1)},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-receipt-wrapper"},(0,n.createElement)("div",{className:"header"},(0,n.createElement)("h2",null,"Payment Receipt"),(0,n.createElement)("span",{className:"close-btn",onClick:()=>h(!1)},"×")),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("div",{className:"receipt-info"},(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Order ID:")," ",q.order_id),(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Date:")," ",new Date(q.order_date).toLocaleDateString()),(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Payment Method:")," ",q.gateway?.title)),(0,n.createElement)("div",{className:"receipt-items"},(0,n.createElement)("h3",null,"Items"),(0,n.createElement)("table",null,(0,n.createElement)("thead",null,(0,n.createElement)("tr",null,(0,n.createElement)("th",null,"Product"),(0,n.createElement)("th",null,"Qty"),(0,n.createElement)("th",null,"Price"))),(0,n.createElement)("tbody",null,q.line_items?.map(((e,t)=>(0,n.createElement)("tr",{key:t},(0,n.createElement)("td",null,e.name),(0,n.createElement)("td",null,e.quantity),(0,n.createElement)("td",null,se(e.quantity*(e.on_sale?e.sale_price:e.regular_price))))))))),(0,n.createElement)("div",{className:"receipt-totals"},(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Subtotal:"),(0,n.createElement)("span",null,se(q.subtotal||0))),q.taxtotal>0&&(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Tax:"),(0,n.createElement)("span",null,se(q.taxtotal||0))),(0,n.createElement)("div",{className:"total-line final-total"},(0,n.createElement)("span",null,"Total:"),(0,n.createElement)("span",null,se(q.ordertotal||0))),"wepos_cash"===z&&(0,n.createElement)(n.Fragment,null,(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Cash Tendered:"),(0,n.createElement)("span",null,se(parseFloat(q.cashamount||"0")))),(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Change:"),(0,n.createElement)("span",null,se(parseFloat(q.changeamount||"0"))))))),(0,n.createElement)("div",{className:"footer"},(0,n.createElement)("button",{className:"print-btn",onClick:()=>{window.print()}},"Print Receipt"),(0,n.createElement)("button",{className:"new-sale-btn",onClick:()=>{h(!1),ve()}},"New Sale"))))))},Be=()=>(0,n.createElement)(ue,null,(0,n.createElement)(se,{path:"/",element:(0,n.createElement)(qe,null)})),Ie=document.getElementById("wepos-react-app");Ie?(0,o.H)(Ie).render((0,n.createElement)(l().StrictMode,null,(0,n.createElement)((function({basename:e,children:t,window:a}){let n=r.useRef();null==n.current&&(n.current=s({window:a,v5Compat:!0}));let l=n.current,[o,i]=r.useState({action:l.action,location:l.location}),c=r.useCallback((e=>{r.startTransition((()=>i(e)))}),[i]);return r.useLayoutEffect((()=>l.listen(c)),[l,c]),r.createElement(ce,{basename:e,children:t,location:o.location,navigationType:o.action,navigator:l})}),null,(0,n.createElement)(Be,null)))):console.error("WePos: React app container not found")})();
+(()=>{"use strict";var e={69:(e,t)=>{Object.prototype.toString},338:(e,t,a)=>{var n=a(795);t.H=n.createRoot,n.hydrateRoot},795:e=>{e.exports=window.ReactDOM}},t={};function a(n){var r=t[n];if(void 0!==r)return r.exports;var l=t[n]={exports:{}};return e[n](l,l.exports,a),l.exports}a.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return a.d(t,{a:t}),t},a.d=(e,t)=>{for(var n in t)a.o(t,n)&&!a.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},a.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t);const n=window.wp.element,r=window.React;var l=a.n(r),o=a(338),i=(a(69),"popstate");function s(e={}){return function(e,t,a,n={}){let{window:r=document.defaultView,v5Compat:l=!1}=n,o=r.history,s="POP",u=null,h=f();function f(){return(o.state||{idx:null}).idx}function g(){s="POP";let e=f(),t=null==e?null:e-h;h=e,u&&u({action:s,location:E.location,delta:t})}function y(e){return function(e,t=!1){let a="http://localhost";"undefined"!=typeof window&&(a="null"!==window.location.origin?window.location.origin:window.location.href),c(a,"No window.location.(origin|href) available to create URL");let n="string"==typeof e?e:p(e);return n=n.replace(/ $/,"%20"),!t&&n.startsWith("//")&&(n=a+n),new URL(n,a)}(e)}null==h&&(h=0,o.replaceState({...o.state,idx:h},""));let E={get action(){return s},get location(){return e(r,o)},listen(e){if(u)throw new Error("A history only accepts one active listener");return r.addEventListener(i,g),u=e,()=>{r.removeEventListener(i,g),u=null}},createHref:e=>t(r,e),createURL:y,encodeLocation(e){let t=y(e);return{pathname:t.pathname,search:t.search,hash:t.hash}},push:function(e,t){s="PUSH";let n=d(E.location,e,t);a&&a(n,e),h=f()+1;let i=m(n,h),c=E.createHref(n);try{o.pushState(i,"",c)}catch(e){if(e instanceof DOMException&&"DataCloneError"===e.name)throw e;r.location.assign(c)}l&&u&&u({action:s,location:E.location,delta:1})},replace:function(e,t){s="REPLACE";let n=d(E.location,e,t);a&&a(n,e),h=f();let r=m(n,h),i=E.createHref(n);o.replaceState(r,"",i),l&&u&&u({action:s,location:E.location,delta:0})},go:e=>o.go(e)};return E}((function(e,t){let{pathname:a="/",search:n="",hash:r=""}=h(e.location.hash.substring(1));return a.startsWith("/")||a.startsWith(".")||(a="/"+a),d("",{pathname:a,search:n,hash:r},t.state&&t.state.usr||null,t.state&&t.state.key||"default")}),(function(e,t){let a=e.document.querySelector("base"),n="";if(a&&a.getAttribute("href")){let t=e.location.href,a=t.indexOf("#");n=-1===a?t:t.slice(0,a)}return n+"#"+("string"==typeof t?t:p(t))}),(function(e,t){u("/"===e.pathname.charAt(0),`relative pathnames are not supported in hash history.push(${JSON.stringify(t)})`)}),e)}function c(e,t){if(!1===e||null==e)throw new Error(t)}function u(e,t){if(!e){"undefined"!=typeof console&&console.warn(t);try{throw new Error(t)}catch(e){}}}function m(e,t){return{usr:e.state,key:e.key,idx:t}}function d(e,t,a=null,n){return{pathname:"string"==typeof e?e:e.pathname,search:"",hash:"",..."string"==typeof t?h(t):t,state:a,key:t&&t.key||n||Math.random().toString(36).substring(2,10)}}function p({pathname:e="/",search:t="",hash:a=""}){return t&&"?"!==t&&(e+="?"===t.charAt(0)?t:"?"+t),a&&"#"!==a&&(e+="#"===a.charAt(0)?a:"#"+a),e}function h(e){let t={};if(e){let a=e.indexOf("#");a>=0&&(t.hash=e.substring(a),e=e.substring(0,a));let n=e.indexOf("?");n>=0&&(t.search=e.substring(n),e=e.substring(0,n)),e&&(t.pathname=e)}return t}function f(e,t,a="/"){return function(e,t,a,n){let r=P(("string"==typeof t?h(t):t).pathname||"/",a);if(null==r)return null;let l=g(e);!function(e){e.sort(((e,t)=>e.score!==t.score?t.score-e.score:function(e,t){return e.length===t.length&&e.slice(0,-1).every(((e,a)=>e===t[a]))?e[e.length-1]-t[t.length-1]:0}(e.routesMeta.map((e=>e.childrenIndex)),t.routesMeta.map((e=>e.childrenIndex)))))}(l);let o=null;for(let e=0;null==o&&e<l.length;++e){let t=R(r);o=k(l[e],t,n)}return o}(e,t,a,!1)}function g(e,t=[],a=[],n=""){let r=(e,r,l)=>{let o={relativePath:void 0===l?e.path||"":l,caseSensitive:!0===e.caseSensitive,childrenIndex:r,route:e};o.relativePath.startsWith("/")&&(c(o.relativePath.startsWith(n),`Absolute route path "${o.relativePath}" nested under path "${n}" is not valid. An absolute child route path must start with the combined path of all its parent routes.`),o.relativePath=o.relativePath.slice(n.length));let i=D([n,o.relativePath]),s=a.concat(o);e.children&&e.children.length>0&&(c(!0!==e.index,`Index routes must not have child routes. Please remove all child routes from route path "${i}".`),g(e.children,t,s,i)),(null!=e.path||e.index)&&t.push({path:i,score:S(i,e.index),routesMeta:s})};return e.forEach(((e,t)=>{if(""!==e.path&&e.path?.includes("?"))for(let a of y(e.path))r(e,t,a);else r(e,t)})),t}function y(e){let t=e.split("/");if(0===t.length)return[];let[a,...n]=t,r=a.endsWith("?"),l=a.replace(/\?$/,"");if(0===n.length)return r?[l,""]:[l];let o=y(n.join("/")),i=[];return i.push(...o.map((e=>""===e?l:[l,e].join("/")))),r&&i.push(...o),i.map((t=>e.startsWith("/")&&""===t?"/":t))}new WeakMap;var E=/^:[\w-]+$/,v=3,w=2,b=1,N=10,C=-2,_=e=>"*"===e;function S(e,t){let a=e.split("/"),n=a.length;return a.some(_)&&(n+=C),t&&(n+=w),a.filter((e=>!_(e))).reduce(((e,t)=>e+(E.test(t)?v:""===t?b:N)),n)}function k(e,t,a=!1){let{routesMeta:n}=e,r={},l="/",o=[];for(let e=0;e<n.length;++e){let i=n[e],s=e===n.length-1,c="/"===l?t:t.slice(l.length)||"/",u=x({path:i.relativePath,caseSensitive:i.caseSensitive,end:s},c),m=i.route;if(!u&&s&&a&&!n[n.length-1].route.index&&(u=x({path:i.relativePath,caseSensitive:i.caseSensitive,end:!1},c)),!u)return null;Object.assign(r,u.params),o.push({params:r,pathname:D([l,u.pathname]),pathnameBase:F(D([l,u.pathnameBase])),route:m}),"/"!==u.pathnameBase&&(l=D([l,u.pathnameBase]))}return o}function x(e,t){"string"==typeof e&&(e={path:e,caseSensitive:!1,end:!0});let[a,n]=function(e,t=!1,a=!0){u("*"===e||!e.endsWith("*")||e.endsWith("/*"),`Route path "${e}" will be treated as if it were "${e.replace(/\*$/,"/*")}" because the \`*\` character must always follow a \`/\` in the pattern. To get rid of this warning, please change the route path to "${e.replace(/\*$/,"/*")}".`);let n=[],r="^"+e.replace(/\/*\*?$/,"").replace(/^\/*/,"/").replace(/[\\.*+^${}|()[\]]/g,"\\$&").replace(/\/:([\w-]+)(\?)?/g,((e,t,a)=>(n.push({paramName:t,isOptional:null!=a}),a?"/?([^\\/]+)?":"/([^\\/]+)")));return e.endsWith("*")?(n.push({paramName:"*"}),r+="*"===e||"/*"===e?"(.*)$":"(?:\\/(.+)|\\/*)$"):a?r+="\\/*$":""!==e&&"/"!==e&&(r+="(?:(?=\\/|$))"),[new RegExp(r,t?void 0:"i"),n]}(e.path,e.caseSensitive,e.end),r=t.match(a);if(!r)return null;let l=r[0],o=l.replace(/(.)\/+$/,"$1"),i=r.slice(1);return{params:n.reduce(((e,{paramName:t,isOptional:a},n)=>{if("*"===t){let e=i[n]||"";o=l.slice(0,l.length-e.length).replace(/(.)\/+$/,"$1")}const r=i[n];return e[t]=a&&!r?void 0:(r||"").replace(/%2F/g,"/"),e}),{}),pathname:l,pathnameBase:o,pattern:e}}function R(e){try{return e.split("/").map((e=>decodeURIComponent(e).replace(/\//g,"%2F"))).join("/")}catch(t){return u(!1,`The URL path "${e}" could not be decoded because it is a malformed URL segment. This is probably due to a bad percent encoding (${t}).`),e}}function P(e,t){if("/"===t)return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let a=t.endsWith("/")?t.length-1:t.length,n=e.charAt(a);return n&&"/"!==n?null:e.slice(a)||"/"}function T(e,t,a,n){return`Cannot include a '${e}' character in a manually specified \`to.${t}\` field [${JSON.stringify(n)}].  Please separate it out to the \`to.${a}\` field. Alternatively you may provide the full path as a string in <Link to="..."> and the router will parse it for you.`}function $(e){let t=function(e){return e.filter(((e,t)=>0===t||e.route.path&&e.route.path.length>0))}(e);return t.map(((e,a)=>a===t.length-1?e.pathname:e.pathnameBase))}function L(e,t,a,n=!1){let r;"string"==typeof e?r=h(e):(r={...e},c(!r.pathname||!r.pathname.includes("?"),T("?","pathname","search",r)),c(!r.pathname||!r.pathname.includes("#"),T("#","pathname","hash",r)),c(!r.search||!r.search.includes("#"),T("#","search","hash",r)));let l,o=""===e||""===r.pathname,i=o?"/":r.pathname;if(null==i)l=a;else{let e=t.length-1;if(!n&&i.startsWith("..")){let t=i.split("/");for(;".."===t[0];)t.shift(),e-=1;r.pathname=t.join("/")}l=e>=0?t[e]:"/"}let s=function(e,t="/"){let{pathname:a,search:n="",hash:r=""}="string"==typeof e?h(e):e,l=a?a.startsWith("/")?a:function(e,t){let a=t.replace(/\/+$/,"").split("/");return e.split("/").forEach((e=>{".."===e?a.length>1&&a.pop():"."!==e&&a.push(e)})),a.length>1?a.join("/"):"/"}(a,t):t;return{pathname:l,search:A(n),hash:M(r)}}(r,l),u=i&&"/"!==i&&i.endsWith("/"),m=(o||"."===i)&&a.endsWith("/");return s.pathname.endsWith("/")||!u&&!m||(s.pathname+="/"),s}var D=e=>e.join("/").replace(/\/\/+/g,"/"),F=e=>e.replace(/\/+$/,"").replace(/^\/*/,"/"),A=e=>e&&"?"!==e?e.startsWith("?")?e:"?"+e:"",M=e=>e&&"#"!==e?e.startsWith("#")?e:"#"+e:"";function O(e){return null!=e&&"number"==typeof e.status&&"string"==typeof e.statusText&&"boolean"==typeof e.internal&&"data"in e}var I=["POST","PUT","PATCH","DELETE"],W=(new Set(I),["GET",...I]);new Set(W),Symbol("ResetLoaderData");var q=r.createContext(null);q.displayName="DataRouter";var B=r.createContext(null);B.displayName="DataRouterState";var U=r.createContext({isTransitioning:!1});U.displayName="ViewTransition",r.createContext(new Map).displayName="Fetchers",r.createContext(null).displayName="Await";var j=r.createContext(null);j.displayName="Navigation";var H=r.createContext(null);H.displayName="Location";var z=r.createContext({outlet:null,matches:[],isDataRoute:!1});z.displayName="Route";var K=r.createContext(null);function G(){return null!=r.useContext(H)}function Q(){return c(G(),"useLocation() may be used only in the context of a <Router> component."),r.useContext(H).location}K.displayName="RouteError";var V="You should call navigate() in a React.useEffect(), not when your component is first rendered.";function J(e){r.useContext(j).static||r.useLayoutEffect(e)}function Y(){let{isDataRoute:e}=r.useContext(z);return e?function(){let{router:e}=function(e){let t=r.useContext(q);return c(t,re(e)),t}("useNavigate"),t=le("useNavigate"),a=r.useRef(!1);return J((()=>{a.current=!0})),r.useCallback((async(n,r={})=>{u(a.current,V),a.current&&("number"==typeof n?e.navigate(n):await e.navigate(n,{fromRouteId:t,...r}))}),[e,t])}():function(){c(G(),"useNavigate() may be used only in the context of a <Router> component.");let e=r.useContext(q),{basename:t,navigator:a}=r.useContext(j),{matches:n}=r.useContext(z),{pathname:l}=Q(),o=JSON.stringify($(n)),i=r.useRef(!1);return J((()=>{i.current=!0})),r.useCallback(((n,r={})=>{if(u(i.current,V),!i.current)return;if("number"==typeof n)return void a.go(n);let s=L(n,JSON.parse(o),l,"path"===r.relative);null==e&&"/"!==t&&(s.pathname="/"===s.pathname?t:D([t,s.pathname])),(r.replace?a.replace:a.push)(s,r.state,r)}),[t,a,o,l,e])}()}function X(e,{relative:t}={}){let{matches:a}=r.useContext(z),{pathname:n}=Q(),l=JSON.stringify($(a));return r.useMemo((()=>L(e,JSON.parse(l),n,"path"===t)),[e,l,n,t])}function Z(e,t,a,n){c(G(),"useRoutes() may be used only in the context of a <Router> component.");let{navigator:l}=r.useContext(j),{matches:o}=r.useContext(z),i=o[o.length-1],s=i?i.params:{},m=i?i.pathname:"/",d=i?i.pathnameBase:"/",p=i&&i.route;{let e=p&&p.path||"";ie(m,!p||e.endsWith("*")||e.endsWith("*?"),`You rendered descendant <Routes> (or called \`useRoutes()\`) at "${m}" (under <Route path="${e}">) but the parent route path has no trailing "*". This means if you navigate deeper, the parent won't match anymore and therefore the child routes will never render.\n\nPlease change the parent <Route path="${e}"> to <Route path="${"/"===e?"*":`${e}/*`}">.`)}let g,y=Q();if(t){let e="string"==typeof t?h(t):t;c("/"===d||e.pathname?.startsWith(d),`When overriding the location using \`<Routes location>\` or \`useRoutes(routes, location)\`, the location pathname must begin with the portion of the URL pathname that was matched by all parent routes. The current pathname base is "${d}" but pathname "${e.pathname}" was given in the \`location\` prop.`),g=e}else g=y;let E=g.pathname||"/",v=E;if("/"!==d){let e=d.replace(/^\//,"").split("/");v="/"+E.replace(/^\//,"").split("/").slice(e.length).join("/")}let w=f(e,{pathname:v});u(p||null!=w,`No routes matched location "${g.pathname}${g.search}${g.hash}" `),u(null==w||void 0!==w[w.length-1].route.element||void 0!==w[w.length-1].route.Component||void 0!==w[w.length-1].route.lazy,`Matched leaf route at location "${g.pathname}${g.search}${g.hash}" does not have an element or Component. This means it will render an <Outlet /> with a null value by default resulting in an "empty" page.`);let b=function(e,t=[],a=null){if(null==e){if(!a)return null;if(a.errors)e=a.matches;else{if(0!==t.length||a.initialized||!(a.matches.length>0))return null;e=a.matches}}let n=e,l=a?.errors;if(null!=l){let e=n.findIndex((e=>e.route.id&&void 0!==l?.[e.route.id]));c(e>=0,`Could not find a matching route for errors on route IDs: ${Object.keys(l).join(",")}`),n=n.slice(0,Math.min(n.length,e+1))}let o=!1,i=-1;if(a)for(let e=0;e<n.length;e++){let t=n[e];if((t.route.HydrateFallback||t.route.hydrateFallbackElement)&&(i=e),t.route.id){let{loaderData:e,errors:r}=a,l=t.route.loader&&!e.hasOwnProperty(t.route.id)&&(!r||void 0===r[t.route.id]);if(t.route.lazy||l){o=!0,n=i>=0?n.slice(0,i+1):[n[0]];break}}}return n.reduceRight(((e,s,c)=>{let u,m=!1,d=null,p=null;a&&(u=l&&s.route.id?l[s.route.id]:void 0,d=s.route.errorElement||te,o&&(i<0&&0===c?(ie("route-fallback",!1,"No `HydrateFallback` element provided to render during initial hydration"),m=!0,p=null):i===c&&(m=!0,p=s.route.hydrateFallbackElement||null)));let h=t.concat(n.slice(0,c+1)),f=()=>{let t;return t=u?d:m?p:s.route.Component?r.createElement(s.route.Component,null):s.route.element?s.route.element:e,r.createElement(ne,{match:s,routeContext:{outlet:e,matches:h,isDataRoute:null!=a},children:t})};return a&&(s.route.ErrorBoundary||s.route.errorElement||0===c)?r.createElement(ae,{location:a.location,revalidation:a.revalidation,component:d,error:u,children:f(),routeContext:{outlet:null,matches:h,isDataRoute:!0}}):f()}),null)}(w&&w.map((e=>Object.assign({},e,{params:Object.assign({},s,e.params),pathname:D([d,l.encodeLocation?l.encodeLocation(e.pathname).pathname:e.pathname]),pathnameBase:"/"===e.pathnameBase?d:D([d,l.encodeLocation?l.encodeLocation(e.pathnameBase).pathname:e.pathnameBase])}))),o,a,n);return t&&b?r.createElement(H.Provider,{value:{location:{pathname:"/",search:"",hash:"",state:null,key:"default",...g},navigationType:"POP"}},b):b}function ee(){let e=function(){let e=r.useContext(K),t=function(e){let t=r.useContext(B);return c(t,re(e)),t}("useRouteError"),a=le("useRouteError");return void 0!==e?e:t.errors?.[a]}(),t=O(e)?`${e.status} ${e.statusText}`:e instanceof Error?e.message:JSON.stringify(e),a=e instanceof Error?e.stack:null,n="rgba(200,200,200, 0.5)",l={padding:"0.5rem",backgroundColor:n},o={padding:"2px 4px",backgroundColor:n},i=null;return console.error("Error handled by React Router default ErrorBoundary:",e),i=r.createElement(r.Fragment,null,r.createElement("p",null,"💿 Hey developer 👋"),r.createElement("p",null,"You can provide a way better UX than this when your app throws errors by providing your own ",r.createElement("code",{style:o},"ErrorBoundary")," or"," ",r.createElement("code",{style:o},"errorElement")," prop on your route.")),r.createElement(r.Fragment,null,r.createElement("h2",null,"Unexpected Application Error!"),r.createElement("h3",{style:{fontStyle:"italic"}},t),a?r.createElement("pre",{style:l},a):null,i)}r.createContext(null);var te=r.createElement(ee,null),ae=class extends r.Component{constructor(e){super(e),this.state={location:e.location,revalidation:e.revalidation,error:e.error}}static getDerivedStateFromError(e){return{error:e}}static getDerivedStateFromProps(e,t){return t.location!==e.location||"idle"!==t.revalidation&&"idle"===e.revalidation?{error:e.error,location:e.location,revalidation:e.revalidation}:{error:void 0!==e.error?e.error:t.error,location:t.location,revalidation:e.revalidation||t.revalidation}}componentDidCatch(e,t){console.error("React Router caught the following error during render",e,t)}render(){return void 0!==this.state.error?r.createElement(z.Provider,{value:this.props.routeContext},r.createElement(K.Provider,{value:this.state.error,children:this.props.component})):this.props.children}};function ne({routeContext:e,match:t,children:a}){let n=r.useContext(q);return n&&n.static&&n.staticContext&&(t.route.errorElement||t.route.ErrorBoundary)&&(n.staticContext._deepestRenderedBoundaryId=t.route.id),r.createElement(z.Provider,{value:e},a)}function re(e){return`${e} must be used within a data router.  See https://reactrouter.com/en/main/routers/picking-a-router.`}function le(e){let t=function(e){let t=r.useContext(z);return c(t,re(e)),t}(e),a=t.matches[t.matches.length-1];return c(a.route.id,`${e} can only be used on routes that contain a unique "id"`),a.route.id}var oe={};function ie(e,t,a){t||oe[e]||(oe[e]=!0,u(!1,a))}function se(e){c(!1,"A <Route> is only ever to be used as the child of <Routes> element, never rendered directly. Please wrap your <Route> in a <Routes>.")}function ce({basename:e="/",children:t=null,location:a,navigationType:n="POP",navigator:l,static:o=!1}){c(!G(),"You cannot render a <Router> inside another <Router>. You should never have more than one in your app.");let i=e.replace(/^\/*/,"/"),s=r.useMemo((()=>({basename:i,navigator:l,static:o,future:{}})),[i,l,o]);"string"==typeof a&&(a=h(a));let{pathname:m="/",search:d="",hash:p="",state:f=null,key:g="default"}=a,y=r.useMemo((()=>{let e=P(m,i);return null==e?null:{location:{pathname:e,search:d,hash:p,state:f,key:g},navigationType:n}}),[i,m,d,p,f,g,n]);return u(null!=y,`<Router basename="${i}"> is not able to match the URL "${m}${d}${p}" because it does not start with the basename, so the <Router> won't render anything.`),null==y?null:r.createElement(j.Provider,{value:s},r.createElement(H.Provider,{children:t,value:y}))}function ue({children:e,location:t}){return Z(me(e),t)}function me(e,t=[]){let a=[];return r.Children.forEach(e,((e,n)=>{if(!r.isValidElement(e))return;let l=[...t,n];if(e.type===r.Fragment)return void a.push.apply(a,me(e.props.children,l));c(e.type===se,`[${"string"==typeof e.type?e.type:e.type.name}] is not a <Route> component. All component children of <Routes> must be a <Route> or <React.Fragment>`),c(!e.props.index||!e.props.children,"An index route cannot have child routes.");let o={id:e.props.id||l.join("-"),caseSensitive:e.props.caseSensitive,element:e.props.element,Component:e.props.Component,index:e.props.index,path:e.props.path,loader:e.props.loader,action:e.props.action,hydrateFallbackElement:e.props.hydrateFallbackElement,HydrateFallback:e.props.HydrateFallback,errorElement:e.props.errorElement,ErrorBoundary:e.props.ErrorBoundary,hasErrorBoundary:!0===e.props.hasErrorBoundary||null!=e.props.ErrorBoundary||null!=e.props.errorElement,shouldRevalidate:e.props.shouldRevalidate,handle:e.props.handle,lazy:e.props.lazy};e.props.children&&(o.children=me(e.props.children,l)),a.push(o)})),a}r.memo((function({routes:e,future:t,state:a}){return Z(e,void 0,a,t)})),r.Component;var de="get",pe="application/x-www-form-urlencoded";function he(e){return null!=e&&"string"==typeof e.tagName}var fe=null,ge=new Set(["application/x-www-form-urlencoded","multipart/form-data","text/plain"]);function ye(e){return null==e||ge.has(e)?e:(u(!1,`"${e}" is not a valid \`encType\` for \`<Form>\`/\`<fetcher.Form>\` and will default to "${pe}"`),null)}function Ee(e,t){if(!1===e||null==e)throw new Error(t)}function ve(e){return null!=e&&(null==e.href?"preload"===e.rel&&"string"==typeof e.imageSrcSet&&"string"==typeof e.imageSizes:"string"==typeof e.rel&&"string"==typeof e.href)}function we(e,t,a,n,r,l){let o=(e,t)=>!a[t]||e.route.id!==a[t].route.id,i=(e,t)=>a[t].pathname!==e.pathname||a[t].route.path?.endsWith("*")&&a[t].params["*"]!==e.params["*"];return"assets"===l?t.filter(((e,t)=>o(e,t)||i(e,t))):"data"===l?t.filter(((t,l)=>{let s=n.routes[t.route.id];if(!s||!s.hasLoader)return!1;if(o(t,l)||i(t,l))return!0;if(t.route.shouldRevalidate){let n=t.route.shouldRevalidate({currentUrl:new URL(r.pathname+r.search+r.hash,window.origin),currentParams:a[0]?.params||{},nextUrl:new URL(e,window.origin),nextParams:t.params,defaultShouldRevalidate:!0});if("boolean"==typeof n)return n}return!0})):[]}function be(e,t,{includeHydrateFallback:a}={}){return n=e.map((e=>{let n=t.routes[e.route.id];if(!n)return[];let r=[n.module];return n.clientActionModule&&(r=r.concat(n.clientActionModule)),n.clientLoaderModule&&(r=r.concat(n.clientLoaderModule)),a&&n.hydrateFallbackModule&&(r=r.concat(n.hydrateFallbackModule)),n.imports&&(r=r.concat(n.imports)),r})).flat(1),[...new Set(n)];var n}Object.getOwnPropertyNames(Object.prototype).sort().join("\0"),"undefined"!=typeof window?window:"undefined"!=typeof globalThis&&globalThis,Symbol("SingleFetchRedirect");function Ne(){let e=r.useContext(q);return Ee(e,"You must render this element inside a <DataRouterContext.Provider> element"),e}function Ce(){let e=r.useContext(B);return Ee(e,"You must render this element inside a <DataRouterStateContext.Provider> element"),e}r.Component;var _e=r.createContext(void 0);function Se(){let e=r.useContext(_e);return Ee(e,"You must render this element inside a <HydratedRouter> element"),e}function ke(e,t){return a=>{e&&e(a),a.defaultPrevented||t(a)}}function xe({page:e,...t}){let{router:a}=Ne(),n=r.useMemo((()=>f(a.routes,e,a.basename)),[a.routes,e,a.basename]);return n?r.createElement(Pe,{page:e,matches:n,...t}):null}function Re(e){let{manifest:t,routeModules:a}=Se(),[n,l]=r.useState([]);return r.useEffect((()=>{let n=!1;return async function(e,t,a){return function(e,t){let a=new Set,n=new Set(t);return e.reduce(((e,r)=>{if(t&&(null==(l=r)||"string"!=typeof l.page)&&"script"===r.as&&r.href&&n.has(r.href))return e;var l;let o=JSON.stringify(function(e){let t={},a=Object.keys(e).sort();for(let n of a)t[n]=e[n];return t}(r));return a.has(o)||(a.add(o),e.push({key:o,link:r})),e}),[])}((await Promise.all(e.map((async e=>{let n=t.routes[e.route.id];if(n){let e=await async function(e,t){if(e.id in t)return t[e.id];try{let a=await import(e.module);return t[e.id]=a,a}catch(t){return console.error(`Error loading route module \`${e.module}\`, reloading page...`),console.error(t),window.__reactRouterContext&&window.__reactRouterContext.isSpaMode,window.location.reload(),new Promise((()=>{}))}}(n,a);return e.links?e.links():[]}return[]})))).flat(1).filter(ve).filter((e=>"stylesheet"===e.rel||"preload"===e.rel)).map((e=>"stylesheet"===e.rel?{...e,rel:"prefetch",as:"style"}:{...e,rel:"prefetch"})))}(e,t,a).then((e=>{n||l(e)})),()=>{n=!0}}),[e,t,a]),n}function Pe({page:e,matches:t,...a}){let n=Q(),{manifest:l,routeModules:o}=Se(),{basename:i}=Ne(),{loaderData:s,matches:c}=Ce(),u=r.useMemo((()=>we(e,t,c,l,n,"data")),[e,t,c,l,n]),m=r.useMemo((()=>we(e,t,c,l,n,"assets")),[e,t,c,l,n]),d=r.useMemo((()=>{if(e===n.pathname+n.search+n.hash)return[];let a=new Set,r=!1;if(t.forEach((e=>{let t=l.routes[e.route.id];t&&t.hasLoader&&(!u.some((t=>t.route.id===e.route.id))&&e.route.id in s&&o[e.route.id]?.shouldRevalidate||t.hasClientLoader?r=!0:a.add(e.route.id))})),0===a.size)return[];let c=function(e,t){let a="string"==typeof e?new URL(e,"undefined"==typeof window?"server://singlefetch/":window.location.origin):e;return"/"===a.pathname?a.pathname="_root.data":t&&"/"===P(a.pathname,t)?a.pathname=`${t.replace(/\/$/,"")}/_root.data`:a.pathname=`${a.pathname.replace(/\/$/,"")}.data`,a}(e,i);return r&&a.size>0&&c.searchParams.set("_routes",t.filter((e=>a.has(e.route.id))).map((e=>e.route.id)).join(",")),[c.pathname+c.search]}),[i,s,n,l,u,t,e,o]),p=r.useMemo((()=>be(m,l)),[m,l]),h=Re(m);return r.createElement(r.Fragment,null,d.map((e=>r.createElement("link",{key:e,rel:"prefetch",as:"fetch",href:e,...a}))),p.map((e=>r.createElement("link",{key:e,rel:"modulepreload",href:e,...a}))),h.map((({key:e,link:t})=>r.createElement("link",{key:e,...t}))))}_e.displayName="FrameworkContext";function Te(...e){return t=>{e.forEach((e=>{"function"==typeof e?e(t):null!=e&&(e.current=t)}))}}var $e="undefined"!=typeof window&&void 0!==window.document&&void 0!==window.document.createElement;try{$e&&(window.__reactRouterVersion="7.6.1")}catch(e){}var Le=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,De=r.forwardRef((function({onClick:e,discover:t="render",prefetch:a="none",relative:n,reloadDocument:l,replace:o,state:i,target:s,to:m,preventScrollReset:d,viewTransition:h,...f},g){let y,{basename:E}=r.useContext(j),v="string"==typeof m&&Le.test(m),w=!1;if("string"==typeof m&&v&&(y=m,$e))try{let e=new URL(window.location.href),t=m.startsWith("//")?new URL(e.protocol+m):new URL(m),a=P(t.pathname,E);t.origin===e.origin&&null!=a?m=a+t.search+t.hash:w=!0}catch(e){u(!1,`<Link to="${m}"> contains an invalid URL which will probably break when clicked - please update to a valid URL path.`)}let b=function(e,{relative:t}={}){c(G(),"useHref() may be used only in the context of a <Router> component.");let{basename:a,navigator:n}=r.useContext(j),{hash:l,pathname:o,search:i}=X(e,{relative:t}),s=o;return"/"!==a&&(s="/"===o?a:D([a,o])),n.createHref({pathname:s,search:i,hash:l})}(m,{relative:n}),[N,C,_]=function(e,t){let a=r.useContext(_e),[n,l]=r.useState(!1),[o,i]=r.useState(!1),{onFocus:s,onBlur:c,onMouseEnter:u,onMouseLeave:m,onTouchStart:d}=t,p=r.useRef(null);r.useEffect((()=>{if("render"===e&&i(!0),"viewport"===e){let e=new IntersectionObserver((e=>{e.forEach((e=>{i(e.isIntersecting)}))}),{threshold:.5});return p.current&&e.observe(p.current),()=>{e.disconnect()}}}),[e]),r.useEffect((()=>{if(n){let e=setTimeout((()=>{i(!0)}),100);return()=>{clearTimeout(e)}}}),[n]);let h=()=>{l(!0)},f=()=>{l(!1),i(!1)};return a?"intent"!==e?[o,p,{}]:[o,p,{onFocus:ke(s,h),onBlur:ke(c,f),onMouseEnter:ke(u,h),onMouseLeave:ke(m,f),onTouchStart:ke(d,h)}]:[!1,p,{}]}(a,f),S=function(e,{target:t,replace:a,state:n,preventScrollReset:l,relative:o,viewTransition:i}={}){let s=Y(),c=Q(),u=X(e,{relative:o});return r.useCallback((r=>{if(function(e,t){return!(0!==e.button||t&&"_self"!==t||function(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}(e))}(r,t)){r.preventDefault();let t=void 0!==a?a:p(c)===p(u);s(e,{replace:t,state:n,preventScrollReset:l,relative:o,viewTransition:i})}}),[c,s,u,a,n,t,e,l,o,i])}(m,{replace:o,state:i,target:s,preventScrollReset:d,relative:n,viewTransition:h}),k=r.createElement("a",{...f,..._,href:y||b,onClick:w||l?e:function(t){e&&e(t),t.defaultPrevented||S(t)},ref:Te(g,C),target:s,"data-discover":v||"render"!==t?void 0:"true"});return N&&!v?r.createElement(r.Fragment,null,k,r.createElement(xe,{page:b})):k}));function Fe(e){let t=r.useContext(q);return c(t,function(e){return`${e} must be used within a data router.  See https://reactrouter.com/en/main/routers/picking-a-router.`}(e)),t}De.displayName="Link",r.forwardRef((function({"aria-current":e="page",caseSensitive:t=!1,className:a="",end:n=!1,style:l,to:o,viewTransition:i,children:s,...u},m){let d=X(o,{relative:u.relative}),p=Q(),h=r.useContext(B),{navigator:f,basename:g}=r.useContext(j),y=null!=h&&function(e,t={}){let a=r.useContext(U);c(null!=a,"`useViewTransitionState` must be used within `react-router-dom`'s `RouterProvider`.  Did you accidentally import `RouterProvider` from `react-router`?");let{basename:n}=Fe("useViewTransitionState"),l=X(e,{relative:t.relative});if(!a.isTransitioning)return!1;let o=P(a.currentLocation.pathname,n)||a.currentLocation.pathname,i=P(a.nextLocation.pathname,n)||a.nextLocation.pathname;return null!=x(l.pathname,i)||null!=x(l.pathname,o)}(d)&&!0===i,E=f.encodeLocation?f.encodeLocation(d).pathname:d.pathname,v=p.pathname,w=h&&h.navigation&&h.navigation.location?h.navigation.location.pathname:null;t||(v=v.toLowerCase(),w=w?w.toLowerCase():null,E=E.toLowerCase()),w&&g&&(w=P(w,g)||w);const b="/"!==E&&E.endsWith("/")?E.length-1:E.length;let N,C=v===E||!n&&v.startsWith(E)&&"/"===v.charAt(b),_=null!=w&&(w===E||!n&&w.startsWith(E)&&"/"===w.charAt(E.length)),S={isActive:C,isPending:_,isTransitioning:y},k=C?e:void 0;N="function"==typeof a?a(S):[a,C?"active":null,_?"pending":null,y?"transitioning":null].filter(Boolean).join(" ");let R="function"==typeof l?l(S):l;return r.createElement(De,{...u,"aria-current":k,className:N,ref:m,style:R,to:o,viewTransition:i},"function"==typeof s?s(S):s)})).displayName="NavLink",r.forwardRef((({discover:e="render",fetcherKey:t,navigate:a,reloadDocument:n,replace:l,state:o,method:i=de,action:s,onSubmit:u,relative:m,preventScrollReset:d,viewTransition:h,...f},g)=>{let y=function(){let{router:e}=Fe("useSubmit"),{basename:t}=r.useContext(j),a=le("useRouteId");return r.useCallback((async(n,r={})=>{let{action:l,method:o,encType:i,formData:s,body:c}=function(e,t){let a,n,r,l,o;if(he(i=e)&&"form"===i.tagName.toLowerCase()){let o=e.getAttribute("action");n=o?P(o,t):null,a=e.getAttribute("method")||de,r=ye(e.getAttribute("enctype"))||pe,l=new FormData(e)}else if(function(e){return he(e)&&"button"===e.tagName.toLowerCase()}(e)||function(e){return he(e)&&"input"===e.tagName.toLowerCase()}(e)&&("submit"===e.type||"image"===e.type)){let o=e.form;if(null==o)throw new Error('Cannot submit a <button> or <input type="submit"> without a <form>');let i=e.getAttribute("formaction")||o.getAttribute("action");if(n=i?P(i,t):null,a=e.getAttribute("formmethod")||o.getAttribute("method")||de,r=ye(e.getAttribute("formenctype"))||ye(o.getAttribute("enctype"))||pe,l=new FormData(o,e),!function(){if(null===fe)try{new FormData(document.createElement("form"),0),fe=!1}catch(e){fe=!0}return fe}()){let{name:t,type:a,value:n}=e;if("image"===a){let e=t?`${t}.`:"";l.append(`${e}x`,"0"),l.append(`${e}y`,"0")}else t&&l.append(t,n)}}else{if(he(e))throw new Error('Cannot submit element that is not <form>, <button>, or <input type="submit|image">');a=de,n=null,r=pe,o=e}var i;return l&&"text/plain"===r&&(o=l,l=void 0),{action:n,method:a.toLowerCase(),encType:r,formData:l,body:o}}(n,t);if(!1===r.navigate){let t=r.fetcherKey||Me();await e.fetch(t,a,r.action||l,{preventScrollReset:r.preventScrollReset,formData:s,body:c,formMethod:r.method||o,formEncType:r.encType||i,flushSync:r.flushSync})}else await e.navigate(r.action||l,{preventScrollReset:r.preventScrollReset,formData:s,body:c,formMethod:r.method||o,formEncType:r.encType||i,replace:r.replace,state:r.state,fromRouteId:a,flushSync:r.flushSync,viewTransition:r.viewTransition})}),[e,t,a])}(),E=function(e,{relative:t}={}){let{basename:a}=r.useContext(j),n=r.useContext(z);c(n,"useFormAction must be used inside a RouteContext");let[l]=n.matches.slice(-1),o={...X(e||".",{relative:t})},i=Q();if(null==e){o.search=i.search;let e=new URLSearchParams(o.search),t=e.getAll("index");if(t.some((e=>""===e))){e.delete("index"),t.filter((e=>e)).forEach((t=>e.append("index",t)));let a=e.toString();o.search=a?`?${a}`:""}}return e&&"."!==e||!l.route.index||(o.search=o.search?o.search.replace(/^\?/,"?index&"):"?index"),"/"!==a&&(o.pathname="/"===o.pathname?a:D([a,o.pathname])),p(o)}(s,{relative:m}),v="get"===i.toLowerCase()?"get":"post",w="string"==typeof s&&Le.test(s);return r.createElement("form",{ref:g,method:v,action:E,onSubmit:n?u:e=>{if(u&&u(e),e.defaultPrevented)return;e.preventDefault();let n=e.nativeEvent.submitter,r=n?.getAttribute("formmethod")||i;y(n||e.currentTarget,{fetcherKey:t,method:r,navigate:a,replace:l,state:o,relative:m,preventScrollReset:d,viewTransition:h})},...f,"data-discover":w||"render"!==e?void 0:"true"})})).displayName="Form";var Ae=0,Me=()=>`__${String(++Ae)}__`;const Oe=window.wp.apiFetch;var Ie=a.n(Oe);const We=e=>{const t="number"==typeof e?e:parseFloat(String(e||0));return isNaN(t)?"$0.00":`$${t.toFixed(2)}`},qe=e=>!e.manage_stock||"instock"===e.stock_status&&e.stock_quantity>0,Be=e=>e.images.length>0?e.images[0].woocommerce_thumbnail:window.wepos?.placeholder_image||"",Ue=(e,t)=>e.length>t?e.substring(0,t)+"...":e,je=e=>parseFloat(e.replace(/[^\d.-]/g,""))||0,He=(e,t)=>{try{if("undefined"==typeof localStorage)return t;const a=localStorage.getItem(e);return a?JSON.parse(a):t}catch(a){return console.error(`Error reading from localStorage for key "${e}":`,a),t}},ze=(e,t)=>{try{"undefined"!=typeof localStorage&&localStorage.setItem(e,JSON.stringify(t))}catch(t){console.error(`Error saving to localStorage for key "${e}":`,t)}},Ke=({products:e,productView:t,productLoading:a,onAddToCart:r,formatPrice:l,hasStock:o,getProductImage:i,truncateTitle:s,itemsWrapperRef:c})=>(0,n.createElement)("div",{className:`items-wrapper ${t}`,ref:c},a?(0,n.createElement)("div",{className:"product-loading"},(0,n.createElement)("div",{className:"spinner spinner-loading"})):(0,n.createElement)(n.Fragment,null,e.length>0?e.map((e=>(0,n.createElement)("div",{key:e.id,className:"item"},"simple"===e.type&&(0,n.createElement)("div",{className:"item-wrap "+(o(e)?"":"disabled"),onClick:()=>r(e)},(0,n.createElement)("div",{className:"img"},(0,n.createElement)("img",{src:i(e),alt:e.name})),(0,n.createElement)("div",{className:"title"},"grid"===t?s(e.name,20):(0,n.createElement)("div",null,(0,n.createElement)("div",{className:"product-name"},e.name),(0,n.createElement)("ul",{className:"meta"},e.sku&&(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"label"},"SKU:"),(0,n.createElement)("span",{className:"value"},e.sku)),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"label"},"Price:"),(0,n.createElement)("span",{className:"value",dangerouslySetInnerHTML:{__html:e.price_html}}))))),(0,n.createElement)("span",{className:`add-product-icon flaticon-add ${t}`}))))):(0,n.createElement)("div",{className:"no-product-found"},(0,n.createElement)("img",{src:`${window.wepos?.assets_url}/images/no-product.png`,alt:"",width:"120px"}),(0,n.createElement)("p",null,"No Product Found")))),Ge=({cartData:e,orderData:t,settings:a,showQuickMenu:r,onShowQuickMenuToggle:o,onUpdateCartItem:i,onRemoveItem:s,onEmptyCart:c,onShowHelp:u,onInitPayment:m,formatPrice:d,getSubtotal:p,getTotalTax:h,getTotal:f})=>{const g=(e,t)=>{i(t,{editQuantity:!e.editQuantity})};return(0,n.createElement)("div",{className:"content-cart"},(0,n.createElement)("div",{className:"top-panel"},(0,n.createElement)("div",{className:"customer-search-box"},(0,n.createElement)("input",{type:"text",id:"customer-search",placeholder:"Walk-in Customer"}),(0,n.createElement)("span",{className:"add-new-customer flaticon-add"})),(0,n.createElement)("div",{className:"action"},(0,n.createElement)("div",{className:"more-options"},(0,n.createElement)("button",{className:"wepos-button",onClick:()=>o(!r)},(0,n.createElement)("span",{className:"more-icon flaticon-more"})),r&&(0,n.createElement)("div",{className:"wepos-dropdown-menu"},(0,n.createElement)("ul",null,(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:c},"Empty Cart")),(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:u},"Help")),(0,n.createElement)("li",{className:"divider"}),(0,n.createElement)("li",null,(0,n.createElement)("a",{href:"#",onClick:()=>window.location.href=window.wepos?.logout_url},"Logout"))))))),a.wepos_general&&(0,n.createElement)("div",{className:"cart-panel"},(0,n.createElement)("div",{className:"cart-content"},(0,n.createElement)("table",{className:"cart-table"},(0,n.createElement)("thead",null,(0,n.createElement)("tr",null,(0,n.createElement)("th",{style:{width:"65%"}},"Product"),(0,n.createElement)("th",{style:{width:"15%"}},"Qty"),(0,n.createElement)("th",{style:{width:"30%"}},"Price"),(0,n.createElement)("th",null),(0,n.createElement)("th",null))),(0,n.createElement)("tbody",null,e.line_items.length>0?e.line_items.map(((e,t)=>(0,n.createElement)(l().Fragment,{key:e.id},(0,n.createElement)("tr",null,(0,n.createElement)("td",{className:"name",onClick:()=>g(e,t)},e.name),(0,n.createElement)("td",{className:"qty",onClick:()=>g(e,t)},e.quantity),(0,n.createElement)("td",{className:"price",onClick:()=>g(e,t)},e.on_sale?(0,n.createElement)(n.Fragment,null,(0,n.createElement)("span",{className:"sale-price"},d(e.quantity*e.sale_price)),(0,n.createElement)("span",{className:"regular-price"},d(e.quantity*e.regular_price))):(0,n.createElement)("span",{className:"sale-price"},d(e.quantity*e.regular_price))),(0,n.createElement)("td",{className:"action"},(0,n.createElement)("span",{className:"flaticon-right-arrow "+(e.editQuantity?"open":""),onClick:()=>g(e,t)})),(0,n.createElement)("td",{className:"remove"},(0,n.createElement)("span",{className:"flaticon-cancel-music",onClick:()=>s(t)}))),e.editQuantity&&(0,n.createElement)("tr",{className:"update-quantity-wrap"},(0,n.createElement)("td",{colSpan:5},(0,n.createElement)("span",{className:"qty"},"Quantity"),(0,n.createElement)("span",{className:"qty-number"},(0,n.createElement)("input",{type:"number",min:"1",step:"1",value:e.quantity,onChange:e=>{i(t,{quantity:parseInt(e.target.value)||1})}})),(0,n.createElement)("span",{className:"qty-action"},(0,n.createElement)("a",{href:"#",className:"add",onClick:a=>{a.preventDefault(),((e,t)=>{i(t,{quantity:e.quantity+1})})(e,t)}},"+"),(0,n.createElement)("a",{href:"#",className:"minus",onClick:a=>{a.preventDefault(),((e,t)=>{e.quantity>1&&i(t,{quantity:e.quantity-1})})(e,t)}},"-"))))))):(0,n.createElement)("tr",{className:"no-item"},(0,n.createElement)("td",{colSpan:5},(0,n.createElement)("img",{src:`${window.wepos?.assets_url}/images/empty-cart.png`,alt:"",width:"120px"}),(0,n.createElement)("p",null,"Empty Cart")))))),(0,n.createElement)("div",{className:"cart-calculation"},(0,n.createElement)("table",{className:"cart-total-table"},(0,n.createElement)("tbody",null,(0,n.createElement)("tr",{className:"cart-meta-data"},(0,n.createElement)("td",{className:"label"},"Subtotal","incl"===a.woo_tax?.wc_tax_display_cart&&h()>0&&(0,n.createElement)("span",{className:"name"},"Including Tax")),(0,n.createElement)("td",{className:"price"},d(p())),(0,n.createElement)("td",{className:"action"})),h()>0&&(0,n.createElement)("tr",{className:"tax"},(0,n.createElement)("td",{className:"label"},"incl"===a.woo_tax?.wc_tax_display_cart?"Fee Tax":"Tax"),(0,n.createElement)("td",{className:"price"},d(h())),(0,n.createElement)("td",{className:"action"})),(0,n.createElement)("tr",{className:"pay-now",onClick:m},(0,n.createElement)("td",null,"Pay Now"),(0,n.createElement)("td",{className:"amount"},d(f())),(0,n.createElement)("td",{className:"icon"},(0,n.createElement)("span",{className:"flaticon-right-arrow"}))))))))},Qe=({show:e,cartData:t,availableGateways:a,selectedGateway:r,cashAmount:l,ableToProcess:o,onGatewayChange:i,onCashAmountChange:s,onBackToSale:c,onProcessPayment:u,formatPrice:m,getTotal:d,changeAmount:p,cashAmountRef:h})=>e?(0,n.createElement)("div",{className:"modal-overlay",onClick:c},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-checkout-wrapper"},(0,n.createElement)("div",{className:"left-content"},(0,n.createElement)("div",{className:"header"},"Sale Summary"),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("table",{className:"sale-summary-cart"},(0,n.createElement)("tbody",null,t.line_items.map((e=>(0,n.createElement)("tr",{key:e.id},(0,n.createElement)("td",{className:"name"},e.name),(0,n.createElement)("td",{className:"quantity"},e.quantity),(0,n.createElement)("td",{className:"price"},e.on_sale?(0,n.createElement)(n.Fragment,null,(0,n.createElement)("span",{className:"sale-price"},m(e.quantity*e.sale_price)),(0,n.createElement)("span",{className:"regular-price"},m(e.quantity*e.regular_price))):(0,n.createElement)("span",{className:"sale-price"},m(e.quantity*e.regular_price)))))))))),(0,n.createElement)("div",{className:"right-content"},(0,n.createElement)("div",{className:"header"},(0,n.createElement)("h2",null,"Pay"),(0,n.createElement)("span",{className:"pay-amount"},m(d()))),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("div",{className:"payment-gateway"},a.length>0?a.map((e=>(0,n.createElement)("label",{key:e.id},(0,n.createElement)("input",{type:"radio",name:"gateway",value:e.id,checked:r===e.id,onChange:e=>i(e.target.value)}),(0,n.createElement)("span",{className:`gateway gateway-${e.id}`},e.title)))):(0,n.createElement)("p",null,"No gateway found")),"wepos_cash"===r&&(0,n.createElement)("div",{className:"payment-option"},(0,n.createElement)("div",{className:"payment-amount"},(0,n.createElement)("div",{className:"input-part"},(0,n.createElement)("div",{className:"input-wrap"},(0,n.createElement)("p",null,"Cash"),(0,n.createElement)("div",{className:"input-addon"},(0,n.createElement)("span",{className:"currency"},"$"),(0,n.createElement)("input",{id:"input-cash-amount",type:"text",value:l,onChange:e=>s(e.target.value),ref:h})))),(0,n.createElement)("div",{className:"change-money"},(0,n.createElement)("p",null,"Change money: ",m(p)))))),(0,n.createElement)("div",{className:"footer"},(0,n.createElement)("a",{href:"#",className:"back-btn",onClick:e=>{e.preventDefault(),c()}},"Back to Sale"),(0,n.createElement)("button",{className:"process-checkout-btn "+(o?"":"disabled"),onClick:u,disabled:!o},"Process Payment")))))):null,Ve=({show:e,printdata:t,selectedGateway:a,onClose:r,onNewSale:l,formatPrice:o})=>e?(0,n.createElement)("div",{className:"modal-overlay",onClick:r},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-receipt-wrapper"},(0,n.createElement)("div",{className:"header"},(0,n.createElement)("h2",null,"Payment Receipt"),(0,n.createElement)("span",{className:"close-btn",onClick:r},"×")),(0,n.createElement)("div",{className:"content"},(0,n.createElement)("div",{className:"receipt-info"},(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Order ID:")," ",t.order_id),(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Date:")," ",t.order_date?new Date(t.order_date).toLocaleDateString():""),(0,n.createElement)("p",null,(0,n.createElement)("strong",null,"Payment Method:")," ",t.gateway?.title)),(0,n.createElement)("div",{className:"receipt-items"},(0,n.createElement)("h3",null,"Items"),(0,n.createElement)("table",null,(0,n.createElement)("thead",null,(0,n.createElement)("tr",null,(0,n.createElement)("th",null,"Product"),(0,n.createElement)("th",null,"Qty"),(0,n.createElement)("th",null,"Price"))),(0,n.createElement)("tbody",null,t.line_items?.map(((e,t)=>(0,n.createElement)("tr",{key:t},(0,n.createElement)("td",null,e.name),(0,n.createElement)("td",null,e.quantity),(0,n.createElement)("td",null,o(e.quantity*(e.on_sale?e.sale_price:e.regular_price))))))))),(0,n.createElement)("div",{className:"receipt-totals"},(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Subtotal:"),(0,n.createElement)("span",null,o(t.subtotal||0))),(t.taxtotal||0)>0&&(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Tax:"),(0,n.createElement)("span",null,o(t.taxtotal||0))),(0,n.createElement)("div",{className:"total-line final-total"},(0,n.createElement)("span",null,"Total:"),(0,n.createElement)("span",null,o(t.ordertotal||0))),"wepos_cash"===a&&(0,n.createElement)(n.Fragment,null,(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Cash Tendered:"),(0,n.createElement)("span",null,o(parseFloat(t.cashamount||"0")))),(0,n.createElement)("div",{className:"total-line"},(0,n.createElement)("span",null,"Change:"),(0,n.createElement)("span",null,o(parseFloat(t.changeamount||"0"))))))),(0,n.createElement)("div",{className:"footer"},(0,n.createElement)("button",{className:"print-btn",onClick:()=>{window.print()}},"Print Receipt"),(0,n.createElement)("button",{className:"new-sale-btn",onClick:()=>{r(),l()}},"New Sale"))))):null,Je=({show:e,onClose:t})=>e?(0,n.createElement)("div",{className:"modal-overlay",onClick:t},(0,n.createElement)("div",{className:"modal-content",onClick:e=>e.stopPropagation()},(0,n.createElement)("div",{className:"wepos-help-wrapper"},(0,n.createElement)("h2",null,"Shortcut Keys"),(0,n.createElement)("ul",null,(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f3")),(0,n.createElement)("span",{className:"title"},"Toggle Product View")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f8")),(0,n.createElement)("span",{className:"title"},"Create New Sale")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"shift+f8")),(0,n.createElement)("span",{className:"title"},"Empty your cart")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"f9")),(0,n.createElement)("span",{className:"title"},"Process Payment")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"ctrl/cmd+?")),(0,n.createElement)("span",{className:"title"},"Show/Close Help")),(0,n.createElement)("li",null,(0,n.createElement)("span",{className:"code"},(0,n.createElement)("code",null,"esc")),(0,n.createElement)("span",{className:"title"},"Close anything")))))):null,Ye=()=>(0,n.createElement)("div",{className:"search-bar"},(0,n.createElement)("div",{className:"search-box"},(0,n.createElement)("input",{type:"text",id:"product-search",placeholder:"Search Product By Name, Barcode, SKU..."}),(0,n.createElement)("span",{className:"search-icon flaticon-search"}))),Xe=({categories:e,selectedCategory:t,onCategoryChange:a})=>(0,n.createElement)("div",{className:"category"},(0,n.createElement)("select",{id:"product-category",value:t?.id||-1,onChange:t=>{const n=parseInt(t.target.value),r=e.find((e=>e.id===n));r&&a(r)}},e.map((e=>(0,n.createElement)("option",{key:e.id,value:e.id},"  ".repeat(e.level),e.name))))),Ze=({productView:e,onToggle:t})=>(0,n.createElement)("div",{className:"toggle-view"},(0,n.createElement)("div",{className:"product-toggle"},(0,n.createElement)("span",{className:"toggle-icon list-view flaticon-menu-button-of-three-horizontal-lines "+("list"===e?"active":""),onClick:t}),(0,n.createElement)("span",{className:"toggle-icon grid-view flaticon-menu "+("grid"===e?"active":""),onClick:t}))),et=()=>{const{products:e,availableGateways:t,settings:a,categories:l,cartData:o,orderData:i,productLoading:s,setCartData:c,setOrderData:u,initializeData:m}=(()=>{const[e,t]=(0,r.useState)([]),[a,n]=(0,r.useState)([]),[l,o]=(0,r.useState)({}),[i,s]=(0,r.useState)([]),[c,u]=(0,r.useState)(1),[m,d]=(0,r.useState)(1),[p,h]=(0,r.useState)(!1),[f,g]=(0,r.useState)((()=>He("cartdata",{line_items:[],fee_lines:[],coupon_lines:[]}))),[y,E]=(0,r.useState)((()=>He("orderdata",{customer_id:0,customer_note:"",payment_method:"",payment_method_title:"",billing:{},shipping:{}}))),v=(0,r.useCallback)((async()=>{if(1===m&&h(!0),c>=m)try{const e=await Ie()({path:`/${window.wepos.rest.posversion}/products?status=publish&per_page=30&page=${m}`}),a=e.filter((e=>!("variable"===e.type&&!e.attributes?.some((e=>!0===e.variation)))));t((e=>[...e,...a])),d((e=>e+1)),h(!1),30===e.length&&setTimeout(v,100)}catch(e){console.error("Error fetching products:",e),h(!1)}else h(!1)}),[m,c]),w=(0,r.useCallback)((async()=>{try{const e=await Ie()({path:`/${window.wepos.rest.posversion}/payment/gateways`});n(e)}catch(e){console.error("Error fetching gateways:",e)}}),[]),b=(0,r.useCallback)((async()=>{try{const e=await Ie()({path:`/${window.wepos.rest.posversion}/settings`});o(e)}catch(e){console.error("Error fetching settings:",e)}}),[]),N=(0,r.useCallback)((async()=>{try{const e=await Ie()({path:`/${window.wepos.rest.wcversion}/products/categories?hide_empty=true&_fields=id,name,parent_id&per_page=100`});e.sort(((e,t)=>e.name.localeCompare(t.name)));const t=(e,a=null,n=0)=>e.filter((e=>e.parent_id===a)).map((a=>[{...a,level:n},...t(e,a.id,n+1)])).flat(),a=t(e);s([{id:-1,level:0,name:"All categories",parent_id:null},...a])}catch(e){console.error("Error fetching categories:",e)}}),[]),C=(0,r.useCallback)((()=>{b(),v(),w(),N()}),[b,v,w,N]);return(0,r.useEffect)((()=>{ze("cartdata",f)}),[f]),(0,r.useEffect)((()=>{ze("orderdata",y)}),[y]),{products:e,availableGateways:a,settings:l,categories:i,cartData:f,orderData:y,productLoading:p,setProducts:t,setCartData:g,setOrderData:E,initializeData:C,fetchProducts:v}})(),{getSubtotal:d,getTotalTax:p,getTotal:h,addToCart:f,updateCartItem:g,removeItem:y,emptyCart:E}=(({cartData:e,setCartData:t,settings:a})=>{const n=(0,r.useCallback)((()=>e.line_items.reduce(((e,t)=>e+(t.on_sale?t.sale_price||0:t.regular_price||0)*(t.quantity||0)),0)),[e.line_items]),l=(0,r.useCallback)((()=>e.line_items.reduce(((e,t)=>e+(t.total_tax||0)),0)),[e.line_items]),o=(0,r.useCallback)((()=>{let t=n()||0;return t+=e.fee_lines.reduce(((e,t)=>e+(t.total||0)),0),t-=e.coupon_lines.reduce(((e,t)=>"discount"===t.type?e+Math.abs(t.total||0):e),0),"incl"!==a.woo_tax?.wc_tax_display_cart&&(t+=l()),t}),[e,a,n,l]),i=(0,r.useCallback)((a=>{if(!qe(a))return void alert("Product is out of stock!");const n=e.line_items.findIndex((e=>e.product_id===a.id));if(-1!==n){const a=[...e.line_items];a[n].quantity+=1,t((e=>({...e,line_items:a})))}else{const e={id:Date.now(),product_id:a.id,name:a.name,quantity:1,type:a.type,on_sale:a.on_sale,sale_price:a.sale_price,regular_price:a.regular_price,attribute:[]};t((t=>({...t,line_items:[...t.line_items,e]})))}}),[e.line_items,t]),s=(0,r.useCallback)(((a,n)=>{const r=[...e.line_items];r[a]={...r[a],...n},t((e=>({...e,line_items:r})))}),[e.line_items,t]),c=(0,r.useCallback)((a=>{const n=e.line_items.filter(((e,t)=>t!==a));t((e=>({...e,line_items:n})))}),[e.line_items,t]),u=(0,r.useCallback)((()=>{t({line_items:[],fee_lines:[],coupon_lines:[]})}),[t]);return{getSubtotal:n,getTotalTax:l,getTotal:o,addToCart:i,updateCartItem:s,removeItem:c,emptyCart:u}})({cartData:o,setCartData:c,settings:a}),[v,w]=(0,r.useState)(!1),[b,N]=(0,r.useState)(!1),[C,_]=(0,r.useState)("grid"),[S,k]=(0,r.useState)(!1),[x,R]=(0,r.useState)(!1),[P,T]=(0,r.useState)(!1),[$,L]=(0,r.useState)(null),[D,F]=(0,r.useState)(""),[A,M]=(0,r.useState)(""),[O,I]=(0,r.useState)({gateway:{id:"",title:""}}),[W,q]=(0,r.useState)(!1),B=(0,r.useRef)(null),U=(0,r.useRef)(null),j=(0,r.useCallback)((()=>{const t=new URLSearchParams(window.location.search).get("category");return null!==t?e.filter((e=>void 0!==e.categories.find((e=>e.id===parseInt(t))))):e}),[e]),H=()=>{_((e=>"grid"===e?"list":"grid"))},z=()=>{E(),u({customer_id:0,customer_note:"",payment_method:"",payment_method_title:"",billing:{},shipping:{}}),R(!1),M(""),N(!1),window.history.pushState({},"",window.location.pathname)},K=()=>{o.line_items.length<=0||(k(!0),t.length>0&&(F(t[0].id),u((e=>({...e,payment_method:t[0].id,payment_method_title:t[0].title})))))},G=()=>{k(!1),w(!1)},Q=(0,r.useCallback)((()=>{const e=je(A)-(h()||0);return e>0?e:0}),[A,h]),V=(0,r.useCallback)((()=>{let e=o.line_items.length>0&&""!==D;return"wepos_cash"===D&&(e=je(A)>=(h()||0)&&e),e}),[o.line_items,D,A,h]);return(0,r.useEffect)((()=>{const e=e=>{switch(e.key){case"F3":e.preventDefault(),H();break;case"F8":e.preventDefault(),e.shiftKey?E():z();break;case"F9":e.preventDefault(),K();break;case"Escape":e.preventDefault(),G();break;case"/":(e.ctrlKey||e.metaKey)&&(e.preventDefault(),w((e=>!e)))}};return window.addEventListener("keydown",e),()=>window.removeEventListener("keydown",e)}),[E]),(0,r.useEffect)((()=>{m()}),[m]),(0,r.useEffect)((()=>{l.length>0&&!$&&L(l[0])}),[l,$]),(0,r.useEffect)((()=>{x&&W&&(setTimeout((()=>{window.confirm("Order successful! Would you like to print the receipt?")&&window.print()}),500),q(!1))}),[x,W]),(0,n.createElement)("div",{id:"wepos-main"},(0,n.createElement)("div",{className:"content-product"},(0,n.createElement)("div",{className:"top-panel wepos-clearfix"},(0,n.createElement)(Ye,null),(0,n.createElement)(Xe,{categories:l,selectedCategory:$,onCategoryChange:L}),(0,n.createElement)(Ze,{productView:C,onToggle:H})),(0,n.createElement)(Ke,{products:j(),productView:C,productLoading:s,onAddToCart:f,formatPrice:We,hasStock:qe,getProductImage:Be,truncateTitle:Ue,itemsWrapperRef:B})),(0,n.createElement)(Ge,{cartData:o,orderData:i,settings:a,showQuickMenu:b,onShowQuickMenuToggle:N,onUpdateCartItem:g,onRemoveItem:y,onEmptyCart:E,onShowHelp:()=>w(!0),onInitPayment:K,formatPrice:We,getSubtotal:d,getTotalTax:p,getTotal:h}),(0,n.createElement)(Je,{show:v,onClose:()=>w(!1)}),(0,n.createElement)(Qe,{show:S,cartData:o,availableGateways:t,selectedGateway:D,cashAmount:A,ableToProcess:V(),onGatewayChange:F,onCashAmountChange:M,onBackToSale:G,onProcessPayment:async()=>{if(V())try{const e=document.querySelector(".wepos-checkout-wrapper");e&&(e.style.opacity="0.6",e.style.pointerEvents="none");const a={billing:i.billing,shipping:i.shipping,line_items:o.line_items.map((e=>({product_id:e.product_id,quantity:e.quantity}))),fee_lines:o.fee_lines,coupon_lines:o.coupon_lines,customer_id:i.customer_id,customer_note:i.customer_note,payment_method:D,payment_method_title:t.find((e=>e.id===D))?.title||"",meta_data:[{key:"_wepos_is_pos_order",value:!0},{key:"_wepos_cash_tendered_amount",value:A.toString()},{key:"_wepos_cash_change_amount",value:Q().toString()}]},n=await Ie()({path:`/${window.wepos.rest.wcversion}/orders`,method:"POST",data:a}),r={};n.line_items.forEach((e=>{r[e.product_id]=e.total_tax}));const l=o.line_items.map((e=>({...e,total_tax:r[e.product_id]||0})));c((e=>({...e,line_items:l}))),"success"===(await Ie()({path:`/${window.wepos.rest.posversion}/payment/process`,method:"POST",data:n})).result&&(I({line_items:o.line_items,fee_lines:o.fee_lines,coupon_lines:o.coupon_lines,subtotal:d(),taxtotal:p(),ordertotal:h(),gateway:{id:n.payment_method,title:n.payment_method_title},order_id:n.number,order_date:n.date_created,cashamount:A.toString(),changeamount:Q().toString()}),k(!1),R(!0),q(!0)),e&&(e.style.opacity="1",e.style.pointerEvents="auto")}catch(e){const t=document.querySelector(".wepos-checkout-wrapper");t&&(t.style.opacity="1",t.style.pointerEvents="auto"),alert(e?.message||"Payment processing failed"),console.error("Payment processing error:",e)}},formatPrice:We,getTotal:h,changeAmount:Q(),cashAmountRef:U}),(0,n.createElement)(Ve,{show:x,printdata:O,selectedGateway:D,onClose:()=>R(!1),onNewSale:z,formatPrice:We}))},tt=()=>(0,n.createElement)(ue,null,(0,n.createElement)(se,{path:"/",element:(0,n.createElement)(et,null)})),at=document.getElementById("wepos-react-app");at?(0,o.H)(at).render((0,n.createElement)(l().StrictMode,null,(0,n.createElement)((function({basename:e,children:t,window:a}){let n=r.useRef();null==n.current&&(n.current=s({window:a,v5Compat:!0}));let l=n.current,[o,i]=r.useState({action:l.action,location:l.location}),c=r.useCallback((e=>{r.startTransition((()=>i(e)))}),[i]);return r.useLayoutEffect((()=>l.listen(c)),[l,c]),r.createElement(ce,{basename:e,children:t,location:o.location,navigationType:o.action,navigator:l})}),null,(0,n.createElement)(tt,null)))):console.error("WePos: React app container not found")})();
