+++ src/frontend/components/POS/CategoryFilter.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { POSCategory } from '../../types';
+
+interface CategoryFilterProps {
+  categories: POSCategory[];
+  selectedCategory: POSCategory | null;
+  onCategoryChange: (category: POSCategory) => void;
+}
+
+const CategoryFilter: React.FC<CategoryFilterProps> = ({
+  categories,
+  selectedCategory,
+  onCategoryChange,
+}) => {
+  return (
+    <div className="category">
+      <select
+        id="product-category"
+        value={selectedCategory?.id || -1}
+        onChange={(e) => {
+          const categoryId = parseInt(e.target.value);
+          const category = categories.find(cat => cat.id === categoryId);
+          if (category) {
+            onCategoryChange(category);
+          }
+        }}
+      >
+        {categories.map(category => (
+          <option key={category.id} value={category.id}>
+            {'  '.repeat(category.level)}{category.name}
+          </option>
+        ))}
+      </select>
+    </div>
+  );
+};
+
+export default CategoryFilter;
